Exclude ioctls from the `general` module.
Ioctl codes are provided in the `ioctl` module, so blocklist them from the
`general` module to avoid duplication.
