+++ gen/ioctl/list.c
@@ -1,0 +1,1971 @@
+// Manually declare things so that we don't pick up any extraneous userspace
+// headers.
+int printf(const char *, ...);
+
+typedef __UINT32_TYPE__ u32;
+typedef __UINT64_TYPE__ u64;
+typedef __INT64_TYPE__ int64_t;
+typedef __UINT64_TYPE__ u_quad_t;
+typedef __SIZE_TYPE__ size_t;
+
+typedef unsigned short u_short;
+typedef unsigned long u_long;
+
+#include <linux/types.h>
+#include <linux/version.h>
+#include <linux/socket.h>
+#include <linux/time.h>
+
+typedef void *caddr_t;
+
+typedef __kernel_ino_t ino_t;
+typedef __kernel_sa_family_t sa_family_t;
+
+struct sockaddr {
+    struct __kernel_sockaddr_storage __storage;
+};
+
+#include <linux/acrn.h>
+#include <linux/agpgart.h>
+#include <linux/am437x-vpfe.h>
+#include <linux/apm_bios.h>
+#include <linux/aspeed-lpc-ctrl.h>
+#include <linux/aspeed-p2a-ctrl.h>
+#include <linux/atm_eni.h>
+#include <linux/atm_idt77105.h>
+#include <linux/atm_he.h>
+#include <linux/atm_tcp.h>
+#include <linux/atm_zatm.h>
+#include <linux/atmarp.h>
+#include <linux/atmbr2684.h>
+#include <linux/atmclip.h>
+#include <linux/atmlec.h>
+#include <linux/atmsvc.h>
+#include <linux/arcfb.h>
+#include <linux/atmdev.h>
+#include <linux/atmmpc.h>
+#include <linux/atm_nicstar.h>
+#include <linux/auto_dev-ioctl.h>
+#include <linux/auto_fs.h>
+#include <linux/blkpg.h>
+#include <linux/blkzoned.h>
+#include <linux/blktrace_api.h>
+#include <linux/bt-bmc.h>
+#include <linux/btrfs.h>
+#include <linux/capi.h>
+#include <linux/cciss_ioctl.h>
+#include <linux/cec.h>
+#include <linux/chio.h>
+#include <linux/cm4000_cs.h>
+#include <linux/coda.h>
+#include <linux/cxl_mem.h>
+#include <linux/dma-buf.h>
+#include <linux/dma-heap.h>
+#include <linux/dm-ioctl.h>
+#include <linux/dn.h>
+#include <linux/fb.h>
+#include <linux/f2fs.h>
+#include <linux/fiemap.h>
+#include <linux/firewire-cdev.h>
+#include <linux/fd.h>
+#include <linux/fpga-dfl.h>
+#include <linux/fsi.h>
+#include <linux/fs.h>
+#include <linux/fsl_hypervisor.h>
+#include <linux/fsl_mc.h>
+#include <linux/fsmap.h>
+#include <linux/fsverity.h>
+#include <linux/fuse.h>
+#include <linux/gpio.h>
+#include <linux/gsmmux.h>
+#include <linux/hiddev.h>
+#include <linux/hidraw.h>
+#include <linux/hpet.h>
+#include <linux/i2o-dev.h>
+#include <linux/i8k.h>
+#include <linux/if_pppox.h>
+#include <linux/if_tun.h>
+#include <linux/inotify.h>
+#include <linux/input.h>
+#include <linux/ioctl.h>
+#include <linux/ipmi.h>
+#include <linux/ipmi_bmc.h>
+#include <linux/ivtvfb.h>
+#include <linux/isst_if.h>
+#include <linux/joystick.h>
+#include <linux/kd.h>
+#include <linux/kcov.h>
+#if !defined(__arm__) && !defined(__powerpc64__) && !defined(__riscv) // various errors
+#include <linux/kvm.h>
+#endif
+#include <linux/lirc.h>
+#include <linux/lp.h>
+#include <linux/matroxfb.h>
+#include <linux/media.h>
+#include <linux/mei.h>
+#include <linux/meye.h>
+#include <linux/mmtimer.h>
+#include <linux/msdos_fs.h>
+#include <linux/mtio.h>
+#include <linux/nbd.h>
+#include <linux/ndctl.h>
+#include <linux/nilfs2_api.h>
+#include <linux/nitro_enclaves.h>
+#include <linux/nvme_ioctl.h>
+#include <linux/nsfs.h>
+#include <linux/nvram.h>
+#include <linux/omap3isp.h>
+#include <linux/pcitest.h>
+#include <linux/perf_event.h>
+#include <linux/phantom.h>
+#include <linux/pktcdvd.h>
+#include <linux/pmu.h>
+#include <linux/pr.h>
+#include <linux/ptp_clock.h>
+#include <linux/ppdev.h>
+#include <linux/ppp-ioctl.h>
+#include <linux/pps.h>
+#include <linux/psp-sev.h>
+#include <linux/radeonfb.h>
+#include <linux/random.h>
+#include <linux/reiserfs_fs.h>
+#include <linux/remoteproc_cdev.h>
+#include <linux/rfkill.h>
+#include <linux/rio_cm_cdev.h>
+#include <linux/rio_mport_cdev.h>
+#include <linux/rpmsg.h>
+#include <linux/rtc.h>
+#include <linux/scif_ioctl.h>
+#include <linux/sed-opal.h>
+#include <linux/serio.h>
+#include <linux/sockios.h>
+#include <linux/sonet.h>
+#include <linux/sonypi.h>
+#include <linux/stm.h>
+#include <linux/suspend_ioctls.h>
+#include <linux/switchtec_ioctl.h>
+#include <linux/sync_file.h>
+#include <linux/synclink.h>
+#include <linux/tee.h>
+#include <linux/termios.h>
+#include <linux/timerfd.h>
+#include <linux/toshiba.h>
+#include <linux/udf_fs_i.h>
+#include <linux/udmabuf.h>
+#include <linux/uinput.h>
+#include <linux/usbdevice_fs.h>
+#if LINUX_VERSION_CODE >= KERNEL_VERSION(4,3,0)
+#include <linux/userfaultfd.h>
+#endif
+#include <linux/uvcvideo.h>
+#include <linux/v4l2-subdev.h>
+#include <linux/vboxguest.h>
+#include <linux/vduse.h>
+#include <linux/vhost.h>
+#include <linux/vfio.h>
+#include <linux/videodev2.h>
+#include <linux/vm_sockets.h>
+#include <linux/vt.h>
+#include <linux/vtpm_proxy.h>
+#include <linux/watch_queue.h>
+#include <linux/watchdog.h>
+#include <linux/wmi.h>
+
+#ifdef __powerpc64__
+typedef __kernel_loff_t loff_t;
+#endif
+
+void entry(const char *s);
+
+#define IOCTL_REQUEST(s) do { \
+    printf("#define " #s " 0x%lxul\n", (unsigned long)(s)); \
+    entry(#s); \
+} while (0)
+
+void list(void) {
+    IOCTL_REQUEST(FIONREAD);
+    IOCTL_REQUEST(FIONBIO);
+    IOCTL_REQUEST(FIOCLEX);
+    IOCTL_REQUEST(FIONCLEX);
+    IOCTL_REQUEST(FIOASYNC);
+    IOCTL_REQUEST(FIOQSIZE);
+
+    IOCTL_REQUEST(TCXONC);
+    IOCTL_REQUEST(TCFLSH);
+    IOCTL_REQUEST(TIOCSCTTY);
+    IOCTL_REQUEST(TIOCSPGRP);
+    IOCTL_REQUEST(TIOCOUTQ);
+    IOCTL_REQUEST(TIOCSTI);
+    IOCTL_REQUEST(TIOCSWINSZ);
+    IOCTL_REQUEST(TIOCMGET);
+    IOCTL_REQUEST(TIOCMBIS);
+    IOCTL_REQUEST(TIOCMBIC);
+    IOCTL_REQUEST(TIOCMSET);
+    IOCTL_REQUEST(TIOCSSOFTCAR);
+#ifdef TIOCTTYGSTRUCT
+    IOCTL_REQUEST(TIOCTTYGSTRUCT);
+#endif
+    IOCTL_REQUEST(TIOCLINUX);
+    IOCTL_REQUEST(TIOCCONS);
+    IOCTL_REQUEST(TIOCSSERIAL);
+    IOCTL_REQUEST(TIOCPKT);
+    IOCTL_REQUEST(TIOCNOTTY);
+    IOCTL_REQUEST(TIOCSETD);
+    IOCTL_REQUEST(TIOCSBRK);
+    IOCTL_REQUEST(TIOCCBRK);
+    IOCTL_REQUEST(TIOCSRS485);
+    IOCTL_REQUEST(TIOCSPTLCK);
+    IOCTL_REQUEST(TIOCSIG);
+    IOCTL_REQUEST(TIOCVHANGUP);
+    IOCTL_REQUEST(TIOCSERCONFIG);
+    IOCTL_REQUEST(TIOCSERGWILD);
+    IOCTL_REQUEST(TIOCSERSWILD);
+    IOCTL_REQUEST(TIOCSLCKTRMIOS);
+    IOCTL_REQUEST(TIOCSERGSTRUCT);
+    IOCTL_REQUEST(TIOCSERGETLSR);
+    IOCTL_REQUEST(TIOCSERGETMULTI);
+    IOCTL_REQUEST(TIOCSERSETMULTI);
+    IOCTL_REQUEST(TIOCMIWAIT);
+    IOCTL_REQUEST(TCGETS);
+    IOCTL_REQUEST(TCGETA);
+    IOCTL_REQUEST(TCSBRK);
+    IOCTL_REQUEST(TCSBRKP);
+    IOCTL_REQUEST(TCSETA);
+    IOCTL_REQUEST(TCSETAF);
+    IOCTL_REQUEST(TCSETAW);
+    IOCTL_REQUEST(TIOCEXCL);
+    IOCTL_REQUEST(TIOCNXCL);
+    IOCTL_REQUEST(TIOCGDEV);
+    IOCTL_REQUEST(TIOCGEXCL);
+    IOCTL_REQUEST(TIOCGICOUNT);
+    IOCTL_REQUEST(TIOCGLCKTRMIOS);
+    IOCTL_REQUEST(TIOCGPGRP);
+    IOCTL_REQUEST(TIOCGPKT);
+    IOCTL_REQUEST(TIOCGPTLCK);
+    IOCTL_REQUEST(TIOCGPTN);
+    IOCTL_REQUEST(TIOCGPTPEER);
+    IOCTL_REQUEST(TIOCGRS485);
+    IOCTL_REQUEST(TIOCGSERIAL);
+    IOCTL_REQUEST(TIOCGSID);
+    IOCTL_REQUEST(TIOCGSOFTCAR);
+    IOCTL_REQUEST(TIOCGWINSZ);
+#ifdef TCGETS2
+    IOCTL_REQUEST(TCGETS2);
+#endif
+#ifdef TCGETX
+    IOCTL_REQUEST(TCGETX);
+#endif
+#ifdef TCSETDTR
+    IOCTL_REQUEST(TCSETDTR);
+#endif
+#ifdef TCSETRTS
+    IOCTL_REQUEST(TCSETRTS);
+#endif
+    IOCTL_REQUEST(TCSETS);
+#ifdef TCSETS2
+    IOCTL_REQUEST(TCSETS2);
+#endif
+    IOCTL_REQUEST(TCSETSF);
+#ifdef TCSETSF2
+    IOCTL_REQUEST(TCSETSF2);
+#endif
+    IOCTL_REQUEST(TCSETSW);
+#ifdef TCSETSW2
+    IOCTL_REQUEST(TCSETSW2);
+#endif
+#ifdef TCSETX
+    IOCTL_REQUEST(TCSETX);
+#endif
+#ifdef TCSETXF
+    IOCTL_REQUEST(TCSETXF);
+#endif
+#ifdef TCSETXW
+    IOCTL_REQUEST(TCSETXW);
+#endif
+#ifdef TIOCGDRAINWAIT
+    IOCTL_REQUEST(TIOCGDRAINWAIT);
+#endif
+#ifdef TIOCGETA
+    IOCTL_REQUEST(TIOCGETA);
+#endif
+#ifdef TIOCGETC
+    IOCTL_REQUEST(TIOCGETC);
+#endif
+    IOCTL_REQUEST(TIOCGETD);
+#ifdef TIOCGETP
+    IOCTL_REQUEST(TIOCGETP);
+#endif
+#ifdef TIOCGLTC
+    IOCTL_REQUEST(TIOCGLTC);
+#endif
+#ifdef TIOCGPPS
+    IOCTL_REQUEST(TIOCGPPS);
+#endif
+#ifdef TIOCGPPSEV
+    IOCTL_REQUEST(TIOCGPPSEV);
+#endif
+    IOCTL_REQUEST(MTIOCGET);
+
+    IOCTL_REQUEST(BLKSSZGET);
+    IOCTL_REQUEST(BLKPBSZGET);
+    IOCTL_REQUEST(BLKROSET);
+    IOCTL_REQUEST(BLKROGET);
+    IOCTL_REQUEST(BLKRRPART);
+    IOCTL_REQUEST(BLKGETSIZE);
+    IOCTL_REQUEST(BLKFLSBUF);
+    IOCTL_REQUEST(BLKRASET);
+    IOCTL_REQUEST(BLKRAGET);
+    IOCTL_REQUEST(BLKFRASET);
+    IOCTL_REQUEST(BLKFRAGET);
+    IOCTL_REQUEST(BLKSECTSET);
+    IOCTL_REQUEST(BLKSECTGET);
+    IOCTL_REQUEST(BLKPG);
+    IOCTL_REQUEST(BLKBSZGET);
+    IOCTL_REQUEST(BLKBSZSET);
+    IOCTL_REQUEST(BLKGETSIZE64);
+    IOCTL_REQUEST(BLKTRACESETUP);
+    IOCTL_REQUEST(BLKTRACESTART);
+    IOCTL_REQUEST(BLKTRACESTOP);
+    IOCTL_REQUEST(BLKTRACETEARDOWN);
+    IOCTL_REQUEST(BLKDISCARD);
+    IOCTL_REQUEST(BLKIOMIN);
+    IOCTL_REQUEST(BLKIOOPT);
+    IOCTL_REQUEST(BLKALIGNOFF);
+    IOCTL_REQUEST(BLKDISCARDZEROES);
+    IOCTL_REQUEST(BLKSECDISCARD);
+    IOCTL_REQUEST(BLKROTATIONAL);
+    IOCTL_REQUEST(BLKZEROOUT);
+
+    // Userfaultfd ioctls.
+    IOCTL_REQUEST(UFFDIO_REGISTER);
+    IOCTL_REQUEST(UFFDIO_UNREGISTER);
+    IOCTL_REQUEST(UFFDIO_WAKE);
+    IOCTL_REQUEST(UFFDIO_COPY);
+    IOCTL_REQUEST(UFFDIO_ZEROPAGE);
+    IOCTL_REQUEST(UFFDIO_WRITEPROTECT);
+    IOCTL_REQUEST(UFFDIO_API);
+
+    IOCTL_REQUEST(NS_GET_USERNS);
+    IOCTL_REQUEST(NS_GET_PARENT);
+    IOCTL_REQUEST(NS_GET_NSTYPE);
+#ifdef NS_GET_OWNED_UID
+    IOCTL_REQUEST(NS_GET_OWNED_UID);
+#endif
+
+    IOCTL_REQUEST(KDGETLED);
+    IOCTL_REQUEST(KDSETLED);
+    IOCTL_REQUEST(KDGKBLED);
+    IOCTL_REQUEST(KDSKBLED);
+    IOCTL_REQUEST(KDGKBTYPE);
+    IOCTL_REQUEST(KDADDIO);
+    IOCTL_REQUEST(KDDELIO);
+    IOCTL_REQUEST(KDENABIO);
+    IOCTL_REQUEST(KDDISABIO);
+    IOCTL_REQUEST(KDSETMODE);
+    IOCTL_REQUEST(KDGETMODE);
+    IOCTL_REQUEST(KDMKTONE);
+    IOCTL_REQUEST(KIOCSOUND);
+    IOCTL_REQUEST(GIO_CMAP);
+    IOCTL_REQUEST(PIO_CMAP);
+    IOCTL_REQUEST(GIO_FONT);
+    IOCTL_REQUEST(GIO_FONTX);
+    IOCTL_REQUEST(PIO_FONT);
+    IOCTL_REQUEST(PIO_FONTX);
+    IOCTL_REQUEST(PIO_FONTRESET);
+    IOCTL_REQUEST(GIO_SCRNMAP);
+    IOCTL_REQUEST(GIO_UNISCRNMAP);
+    IOCTL_REQUEST(PIO_SCRNMAP);
+    IOCTL_REQUEST(PIO_UNISCRNMAP);
+    IOCTL_REQUEST(GIO_UNIMAP);
+    IOCTL_REQUEST(PIO_UNIMAP);
+    IOCTL_REQUEST(PIO_UNIMAPCLR);
+    IOCTL_REQUEST(KDGKBMODE);
+    IOCTL_REQUEST(KDSKBMODE);
+    IOCTL_REQUEST(KDGKBMETA);
+    IOCTL_REQUEST(KDSKBMETA);
+    IOCTL_REQUEST(KDGKBENT);
+    IOCTL_REQUEST(KDSKBENT);
+    IOCTL_REQUEST(KDGKBSENT);
+    IOCTL_REQUEST(KDSKBSENT);
+    IOCTL_REQUEST(KDGKBDIACR);
+    IOCTL_REQUEST(KDGETKEYCODE);
+    IOCTL_REQUEST(KDSETKEYCODE);
+    IOCTL_REQUEST(KDSIGACCEPT);
+    IOCTL_REQUEST(VT_OPENQRY);
+    IOCTL_REQUEST(VT_GETMODE);
+    IOCTL_REQUEST(VT_SETMODE);
+    IOCTL_REQUEST(VT_GETSTATE);
+    IOCTL_REQUEST(VT_RELDISP);
+    IOCTL_REQUEST(VT_ACTIVATE);
+    IOCTL_REQUEST(VT_WAITACTIVE);
+    IOCTL_REQUEST(VT_DISALLOCATE);
+    IOCTL_REQUEST(VT_RESIZE);
+    IOCTL_REQUEST(VT_RESIZEX);
+    IOCTL_REQUEST(FIOSETOWN);
+    IOCTL_REQUEST(SIOCSPGRP);
+    IOCTL_REQUEST(FIOGETOWN);
+    IOCTL_REQUEST(SIOCGPGRP);
+    IOCTL_REQUEST(SIOCATMARK);
+    IOCTL_REQUEST(SIOCGSTAMP);
+    IOCTL_REQUEST(TIOCINQ);
+    IOCTL_REQUEST(SIOCADDRT);
+    IOCTL_REQUEST(SIOCDELRT);
+    IOCTL_REQUEST(SIOCGIFNAME);
+    IOCTL_REQUEST(SIOCSIFLINK);
+    IOCTL_REQUEST(SIOCGIFCONF);
+    IOCTL_REQUEST(SIOCGIFFLAGS);
+    IOCTL_REQUEST(SIOCSIFFLAGS);
+    IOCTL_REQUEST(SIOCGIFADDR);
+    IOCTL_REQUEST(SIOCSIFADDR);
+    IOCTL_REQUEST(SIOCGIFDSTADDR);
+    IOCTL_REQUEST(SIOCSIFDSTADDR);
+    IOCTL_REQUEST(SIOCGIFBRDADDR);
+    IOCTL_REQUEST(SIOCSIFBRDADDR);
+    IOCTL_REQUEST(SIOCGIFNETMASK);
+    IOCTL_REQUEST(SIOCSIFNETMASK);
+    IOCTL_REQUEST(SIOCGIFMETRIC);
+    IOCTL_REQUEST(SIOCSIFMETRIC);
+    IOCTL_REQUEST(SIOCGIFMEM);
+    IOCTL_REQUEST(SIOCSIFMEM);
+    IOCTL_REQUEST(SIOCGIFMTU);
+    IOCTL_REQUEST(SIOCSIFMTU);
+    IOCTL_REQUEST(SIOCSIFHWADDR);
+    IOCTL_REQUEST(SIOCGIFENCAP);
+    IOCTL_REQUEST(SIOCSIFENCAP);
+    IOCTL_REQUEST(SIOCGIFHWADDR);
+    IOCTL_REQUEST(SIOCGIFSLAVE);
+    IOCTL_REQUEST(SIOCSIFSLAVE);
+    IOCTL_REQUEST(SIOCADDMULTI);
+    IOCTL_REQUEST(SIOCDELMULTI);
+    IOCTL_REQUEST(SIOCDARP);
+    IOCTL_REQUEST(SIOCGARP);
+    IOCTL_REQUEST(SIOCSARP);
+    IOCTL_REQUEST(SIOCDRARP);
+    IOCTL_REQUEST(SIOCGRARP);
+    IOCTL_REQUEST(SIOCSRARP);
+    IOCTL_REQUEST(SIOCGIFMAP);
+    IOCTL_REQUEST(SIOCSIFMAP);
+
+    IOCTL_REQUEST(FIBMAP);
+    IOCTL_REQUEST(FIGETBSZ);
+    IOCTL_REQUEST(FIFREEZE);
+    IOCTL_REQUEST(FITHAW);
+    IOCTL_REQUEST(FITRIM);
+    IOCTL_REQUEST(FICLONE);
+    IOCTL_REQUEST(FICLONERANGE);
+    IOCTL_REQUEST(FIDEDUPERANGE);
+
+    IOCTL_REQUEST(FS_IOC_GETFLAGS);
+    IOCTL_REQUEST(FS_IOC_SETFLAGS);
+    IOCTL_REQUEST(FS_IOC_GETVERSION);
+    IOCTL_REQUEST(FS_IOC_SETVERSION);
+    IOCTL_REQUEST(FS_IOC_FIEMAP);
+    IOCTL_REQUEST(FS_IOC32_GETFLAGS);
+    IOCTL_REQUEST(FS_IOC32_SETFLAGS);
+    IOCTL_REQUEST(FS_IOC32_GETVERSION);
+    IOCTL_REQUEST(FS_IOC32_SETVERSION);
+    IOCTL_REQUEST(FS_IOC_FSGETXATTR);
+    IOCTL_REQUEST(FS_IOC_FSSETXATTR);
+    IOCTL_REQUEST(FS_IOC_GETFSLABEL);
+    IOCTL_REQUEST(FS_IOC_SETFSLABEL);
+
+    IOCTL_REQUEST(VIDIOC_SUBDEV_QUERYSTD);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_CLOSEMOUNT);
+    IOCTL_REQUEST(LIRC_SET_SEND_CARRIER);
+    IOCTL_REQUEST(AUTOFS_IOC_PROTOSUBVER);
+    IOCTL_REQUEST(PTP_SYS_OFFSET_PRECISE);
+    IOCTL_REQUEST(FSI_SCOM_WRITE);
+    IOCTL_REQUEST(ATM_GETCIRANGE);
+    IOCTL_REQUEST(DMA_BUF_SET_NAME_B);
+    IOCTL_REQUEST(RIO_CM_EP_GET_LIST_SIZE);
+    IOCTL_REQUEST(TUNSETPERSIST);
+    IOCTL_REQUEST(FS_IOC_GET_ENCRYPTION_POLICY);
+    IOCTL_REQUEST(CEC_RECEIVE);
+    IOCTL_REQUEST(MGSL_IOCGPARAMS);
+    IOCTL_REQUEST(ENI_SETMULT);
+    IOCTL_REQUEST(RIO_GET_EVENT_MASK);
+    IOCTL_REQUEST(LIRC_GET_MAX_TIMEOUT);
+#if !defined(__arm__) && !defined(__powerpc64__) && !defined(__riscv)
+#if 0 // needs `struct kvm_cpuid2`
+    IOCTL_REQUEST(KVM_GET_SUPPORTED_CPUID);
+#endif
+    IOCTL_REQUEST(KVM_SET_BOOT_CPU_ID);
+    IOCTL_REQUEST(KVM_SET_MP_STATE);
+    IOCTL_REQUEST(KVM_GET_MP_STATE);
+    IOCTL_REQUEST(KVM_SET_MEMORY_REGION);
+    IOCTL_REQUEST(KVM_GET_API_VERSION);
+    IOCTL_REQUEST(KVM_S390_GET_SKEYS);
+#if 0 // needs `struct kvm_create_spapr_tce_64`
+    IOCTL_REQUEST(KVM_CREATE_SPAPR_TCE_64);
+#endif
+    IOCTL_REQUEST(KVM_SET_ONE_REG);
+#if 0 // needs `struct kvm_xcrs`
+    IOCTL_REQUEST(KVM_SET_XCRS);
+#endif
+    IOCTL_REQUEST(KVM_IRQ_LINE_STATUS);
+    IOCTL_REQUEST(KVM_SET_USER_MEMORY_REGION);
+    IOCTL_REQUEST(KVM_S390_INTERRUPT);
+    IOCTL_REQUEST(KVM_TPR_ACCESS_REPORTING);
+#if 0 // needs `struct kvm_xsave`
+    IOCTL_REQUEST(KVM_SET_XSAVE);
+#endif
+    IOCTL_REQUEST(KVM_GET_FPU);
+    IOCTL_REQUEST(KVM_ASSIGN_SET_MSIX_ENTRY);
+    IOCTL_REQUEST(KVM_MEMORY_ENCRYPT_REG_REGION);
+    IOCTL_REQUEST(KVM_S390_UCAS_MAP);
+    IOCTL_REQUEST(KVM_ASSIGN_SET_INTX_MASK);
+#if 0 // needs `struct kvm_xen_hvm_config`
+    IOCTL_REQUEST(KVM_XEN_HVM_CONFIG);
+#endif
+#if 0 // needs `struct kvm_x86_mce`
+    IOCTL_REQUEST(KVM_X86_SET_MCE);
+#endif
+#if 0 // needs `struct kvm_vcpu_init`
+    IOCTL_REQUEST(KVM_ARM_PREFERRED_TARGET);
+#endif
+    IOCTL_REQUEST(KVM_MEMORY_ENCRYPT_OP);
+    IOCTL_REQUEST(KVM_CREATE_PIT2);
+#if 0 // needs `struct kvm_rtas_token_args`
+    IOCTL_REQUEST(KVM_PPC_RTAS_DEFINE_TOKEN);
+#endif
+#if 0 // needs `struct kvm_get_htab_fd`
+    IOCTL_REQUEST(KVM_PPC_GET_HTAB_FD);
+#endif
+    IOCTL_REQUEST(KVM_S390_VCPU_FAULT);
+    IOCTL_REQUEST(KVM_REGISTER_COALESCED_MMIO);
+    IOCTL_REQUEST(KVM_SET_VAPIC_ADDR);
+    IOCTL_REQUEST(KVM_CREATE_DEVICE);
+    IOCTL_REQUEST(KVM_GET_IRQCHIP);
+#if 0 // needs `struct kvm_nested_state`
+    IOCTL_REQUEST(KVM_SET_NESTED_STATE);
+#endif
+    IOCTL_REQUEST(KVM_TRANSLATE);
+    IOCTL_REQUEST(KVM_S390_SET_INITIAL_PSW);
+    IOCTL_REQUEST(KVM_X86_SETUP_MCE);
+    IOCTL_REQUEST(KVM_DEASSIGN_PCI_DEVICE);
+#if 0 // needs `struct kvm_nested_state`
+    IOCTL_REQUEST(KVM_GET_NESTED_STATE);
+#endif
+    IOCTL_REQUEST(KVM_XEN_HVM_SET_ATTR);
+    IOCTL_REQUEST(KVM_GET_CLOCK);
+#if 0 // needs `struct kvm_cpuid`
+    IOCTL_REQUEST(KVM_SET_CPUID);
+#endif
+    IOCTL_REQUEST(KVM_SET_FPU);
+    IOCTL_REQUEST(KVM_KVMCLOCK_CTRL);
+#ifdef KVM_GET_SREGS2
+    IOCTL_REQUEST(KVM_GET_SREGS2);
+#endif
+    IOCTL_REQUEST(KVM_PPC_GET_PVINFO);
+    IOCTL_REQUEST(KVM_X86_GET_MCE_CAP_SUPPORTED);
+    IOCTL_REQUEST(KVM_S390_GET_IRQ_STATE);
+#if 0 // needs `struct kvm_xcrs`
+    IOCTL_REQUEST(KVM_GET_XCRS);
+#endif
+    IOCTL_REQUEST(KVM_GET_DEVICE_ATTR);
+    IOCTL_REQUEST(KVM_S390_INITIAL_RESET);
+    IOCTL_REQUEST(KVM_SET_REGS);
+#if 0 // needs `struct kvm_msr_list`
+    IOCTL_REQUEST(KVM_GET_MSR_INDEX_LIST);
+#endif
+#if 0 // needs `struct kvm_msrs`
+    IOCTL_REQUEST(KVM_SET_MSRS);
+#endif
+    IOCTL_REQUEST(KVM_GET_TSC_KHZ);
+#if 0 // needs `struct kvm_ppc_mmuv3_cfg`
+    IOCTL_REQUEST(KVM_PPC_CONFIGURE_V3_MMU);
+#endif
+    IOCTL_REQUEST(KVM_CHECK_EXTENSION);
+    IOCTL_REQUEST(KVM_PPC_RESIZE_HPT_COMMIT);
+    IOCTL_REQUEST(KVM_GET_DIRTY_LOG);
+    IOCTL_REQUEST(KVM_S390_SET_IRQ_STATE);
+    IOCTL_REQUEST(KVM_S390_NORMAL_RESET);
+#if 0 // needs `struct kvm_lapic_state`
+    IOCTL_REQUEST(KVM_SET_LAPIC);
+#endif
+    IOCTL_REQUEST(KVM_SET_SIGNAL_MASK);
+#if 0 // needs `struct kvm_pit_state`
+    IOCTL_REQUEST(KVM_GET_PIT);
+#endif
+#if 0 // needs `struct kvm_arm_copy_mte_tags`
+    IOCTL_REQUEST(KVM_ARM_MTE_COPY_TAGS);
+#endif
+#if 0 // needs `struct kvm_ppc_cpu_char`
+    IOCTL_REQUEST(KVM_PPC_GET_CPU_CHAR);
+#endif
+#if 0 // needs `struct kvm_vcpu_events`
+    IOCTL_REQUEST(KVM_GET_VCPU_EVENTS);
+#endif
+    IOCTL_REQUEST(KVM_XEN_VCPU_SET_ATTR);
+    IOCTL_REQUEST(KVM_SET_GSI_ROUTING);
+    IOCTL_REQUEST(KVM_CREATE_VCPU);
+    IOCTL_REQUEST(KVM_S390_UCAS_UNMAP);
+    IOCTL_REQUEST(KVM_S390_STORE_STATUS);
+    IOCTL_REQUEST(KVM_UNREGISTER_COALESCED_MMIO);
+    IOCTL_REQUEST(KVM_ASSIGN_DEV_IRQ);
+    IOCTL_REQUEST(KVM_S390_PV_COMMAND);
+    IOCTL_REQUEST(KVM_S390_ENABLE_SIE);
+    IOCTL_REQUEST(KVM_CLEAR_DIRTY_LOG);
+    IOCTL_REQUEST(KVM_GET_ONE_REG);
+#if 0 // needs `struct kvm_ppc_rmmu_info`
+    IOCTL_REQUEST(KVM_PPC_GET_RMMU_INFO);
+#endif
+    IOCTL_REQUEST(KVM_ARM_SET_DEVICE_ADDR);
+    IOCTL_REQUEST(KVM_DIRTY_TLB);
+    IOCTL_REQUEST(KVM_NMI);
+    IOCTL_REQUEST(KVM_HYPERV_EVENTFD);
+    IOCTL_REQUEST(KVM_S390_MEM_OP);
+#if 0 // needs `struct kvm_pit_state2`
+    IOCTL_REQUEST(KVM_GET_PIT2);
+#endif
+    IOCTL_REQUEST(KVM_PPC_SVM_OFF);
+    IOCTL_REQUEST(KVM_PPC_RESIZE_HPT_PREPARE);
+#if 0 // needs `struct kvm_cpuid2`
+    IOCTL_REQUEST(KVM_GET_SUPPORTED_HV_CPUID);
+#endif
+    IOCTL_REQUEST(KVM_INTERRUPT);
+#if 0 // needs `struct kvm_create_spapr_tce`
+    IOCTL_REQUEST(KVM_CREATE_SPAPR_TCE);
+#endif
+    IOCTL_REQUEST(KVM_XEN_HVM_GET_ATTR);
+#if 0 // needs `struct kvm_xsave`
+    IOCTL_REQUEST(KVM_GET_XSAVE);
+#endif
+    IOCTL_REQUEST(KVM_GET_SREGS);
+    IOCTL_REQUEST(KVM_SET_IDENTITY_MAP_ADDR);
+#if 0 // needs `struct kvm_msr_list`
+    IOCTL_REQUEST(KVM_GET_MSR_FEATURE_INDEX_LIST);
+#endif
+#if 0 // needs `struct kvm_vcpu_init`
+    IOCTL_REQUEST(KVM_ARM_VCPU_INIT);
+#endif
+#if 0 // needs `struct kvm_cpuid2`
+    IOCTL_REQUEST(KVM_GET_CPUID2);
+#endif
+    IOCTL_REQUEST(KVM_GET_REGS);
+#if 0 // needs `struct kvm_msrs`
+    IOCTL_REQUEST(KVM_GET_MSRS);
+#endif
+    IOCTL_REQUEST(KVM_S390_GET_CMMA_BITS);
+    IOCTL_REQUEST(KVM_CREATE_IRQCHIP);
+    IOCTL_REQUEST(KVM_ASSIGN_PCI_DEVICE);
+    IOCTL_REQUEST(KVM_SET_NR_MMU_PAGES);
+    IOCTL_REQUEST(KVM_IOEVENTFD);
+#if 0 // needs `struct kvm_pit_state`
+    IOCTL_REQUEST(KVM_SET_PIT);
+#endif
+    IOCTL_REQUEST(KVM_RUN);
+    IOCTL_REQUEST(KVM_S390_SET_SKEYS);
+    IOCTL_REQUEST(KVM_SET_DEVICE_ATTR);
+    IOCTL_REQUEST(KVM_DEASSIGN_DEV_IRQ);
+    IOCTL_REQUEST(KVM_SMI);
+    IOCTL_REQUEST(KVM_SET_TSS_ADDR);
+    IOCTL_REQUEST(KVM_GET_NR_MMU_PAGES);
+    IOCTL_REQUEST(KVM_XEN_VCPU_GET_ATTR);
+#if 0 // needs `struct kvm_debugregs`
+    IOCTL_REQUEST(KVM_GET_DEBUGREGS);
+#endif
+    IOCTL_REQUEST(KVM_REINJECT_CONTROL);
+    IOCTL_REQUEST(KVM_HAS_DEVICE_ATTR);
+#if 0 // needs `struct kvm_memory_alias`
+    IOCTL_REQUEST(KVM_SET_MEMORY_ALIAS);
+#endif
+    IOCTL_REQUEST(KVM_IRQFD);
+#if 0 // needs `struct kvm_vcpu_events`
+    IOCTL_REQUEST(KVM_SET_VCPU_EVENTS);
+#endif
+#if 0 // needs `struct kvm_debugregs`
+    IOCTL_REQUEST(KVM_SET_DEBUGREGS);
+#endif
+    IOCTL_REQUEST(KVM_S390_SET_CMMA_BITS);
+#if 0 // needs `struct kvm_lapic_state`
+    IOCTL_REQUEST(KVM_GET_LAPIC);
+#endif
+    IOCTL_REQUEST(KVM_SET_IRQCHIP);
+#ifdef KVM_SET_SREGS2
+    IOCTL_REQUEST(KVM_SET_SREGS2);
+#endif
+    IOCTL_REQUEST(KVM_SIGNAL_MSI);
+    IOCTL_REQUEST(KVM_GET_VCPU_MMAP_SIZE);
+    IOCTL_REQUEST(KVM_CREATE_VM);
+#if 0 // needs `struct kvm_allocate_rma`
+    IOCTL_REQUEST(KVM_ALLOCATE_RMA);
+#endif
+    IOCTL_REQUEST(KVM_S390_CLEAR_RESET);
+    IOCTL_REQUEST(KVM_RESET_DIRTY_RINGS);
+    IOCTL_REQUEST(KVM_ASSIGN_SET_MSIX_NR);
+    IOCTL_REQUEST(KVM_CREATE_PIT);
+    IOCTL_REQUEST(KVM_SET_GUEST_DEBUG);
+    IOCTL_REQUEST(KVM_GET_REG_LIST);
+    IOCTL_REQUEST(KVM_ARM_VCPU_FINALIZE);
+#if 0 // needs `struct kvm_cpuid2`
+    IOCTL_REQUEST(KVM_SET_CPUID2);
+#endif
+#if 0 // needs `struct kvm_pit_state2`
+    IOCTL_REQUEST(KVM_SET_PIT2);
+#endif
+    IOCTL_REQUEST(KVM_SET_CLOCK);
+    IOCTL_REQUEST(KVM_PPC_ALLOCATE_HTAB);
+    IOCTL_REQUEST(KVM_PPC_GET_SMMU_INFO);
+#if 0 // needs `struct kvm_msr_filter`
+    IOCTL_REQUEST(KVM_X86_SET_MSR_FILTER);
+#endif
+    IOCTL_REQUEST(KVM_SET_TSC_KHZ);
+    IOCTL_REQUEST(KVM_S390_IRQ);
+    IOCTL_REQUEST(KVM_ENABLE_CAP);
+    IOCTL_REQUEST(KVM_IRQ_LINE);
+    IOCTL_REQUEST(KVM_SET_SREGS);
+    IOCTL_REQUEST(KVM_MEMORY_ENCRYPT_UNREG_REGION);
+#if 0 // needs `struct kvm_cpuid2`
+    IOCTL_REQUEST(KVM_GET_EMULATED_CPUID);
+#endif
+#if 0 // needs `struct kvm_pmu_event_filter`
+    IOCTL_REQUEST(KVM_SET_PMU_EVENT_FILTER);
+#endif
+#ifdef KVM_GET_STATS_FD
+    IOCTL_REQUEST(KVM_GET_STATS_FD);
+#endif
+#endif
+    IOCTL_REQUEST(USBDEVFS_CLAIMINTERFACE);
+    IOCTL_REQUEST(CHIOMOVE);
+    IOCTL_REQUEST(SONYPI_IOCGBATFLAGS);
+    IOCTL_REQUEST(BTRFS_IOC_SYNC);
+#if 0 // needs `compat_ulong_t`
+    IOCTL_REQUEST(AUTOFS_IOC_SETTIMEOUT32);
+#endif
+    IOCTL_REQUEST(VIDIOC_TRY_FMT);
+    IOCTL_REQUEST(LIRC_SET_REC_MODE);
+    IOCTL_REQUEST(VIDIOC_DQEVENT);
+    IOCTL_REQUEST(RPMSG_DESTROY_EPT_IOCTL);
+    IOCTL_REQUEST(UVCIOC_CTRL_MAP);
+    IOCTL_REQUEST(VHOST_SET_BACKEND_FEATURES);
+    IOCTL_REQUEST(VHOST_VSOCK_SET_GUEST_CID);
+    IOCTL_REQUEST(UI_SET_KEYBIT);
+    IOCTL_REQUEST(LIRC_SET_REC_TIMEOUT);
+    IOCTL_REQUEST(FS_IOC_GET_ENCRYPTION_KEY_STATUS);
+    IOCTL_REQUEST(BTRFS_IOC_TREE_SEARCH_V2);
+    IOCTL_REQUEST(VHOST_SET_VRING_BASE);
+    IOCTL_REQUEST(RIO_ENABLE_DOORBELL_RANGE);
+    IOCTL_REQUEST(VIDIOC_TRY_EXT_CTRLS);
+    IOCTL_REQUEST(LIRC_GET_REC_MODE);
+    IOCTL_REQUEST(PPGETTIME);
+    IOCTL_REQUEST(BTRFS_IOC_RM_DEV);
+    IOCTL_REQUEST(ATM_SETBACKEND);
+    IOCTL_REQUEST(FSL_HV_IOCTL_PARTITION_START);
+    IOCTL_REQUEST(FBIO_WAITEVENT);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_PORT_TO_PFF);
+    IOCTL_REQUEST(NVME_IOCTL_IO_CMD);
+    IOCTL_REQUEST(IPMICTL_RECEIVE_MSG_TRUNC);
+    IOCTL_REQUEST(FDTWADDLE);
+    IOCTL_REQUEST(NVME_IOCTL_SUBMIT_IO);
+    IOCTL_REQUEST(NILFS_IOCTL_SYNC);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_DV_TIMINGS);
+    IOCTL_REQUEST(ASPEED_LPC_CTRL_IOCTL_GET_SIZE);
+    IOCTL_REQUEST(DM_DEV_STATUS);
+    IOCTL_REQUEST(TEE_IOC_CLOSE_SESSION);
+    IOCTL_REQUEST(NS_GETPSTAT);
+    IOCTL_REQUEST(UI_SET_PROPBIT);
+    IOCTL_REQUEST(TUNSETFILTEREBPF);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_COMPTAG_SET);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_VERSION);
+    IOCTL_REQUEST(WDIOC_SETOPTIONS);
+    IOCTL_REQUEST(VHOST_SCSI_SET_ENDPOINT);
+    IOCTL_REQUEST(MGSL_IOCGTXIDLE);
+    IOCTL_REQUEST(ATM_ADDLECSADDR);
+    IOCTL_REQUEST(FSL_HV_IOCTL_GETPROP);
+    IOCTL_REQUEST(FDGETPRM);
+    IOCTL_REQUEST(HIDIOCAPPLICATION);
+    IOCTL_REQUEST(ENI_MEMDUMP);
+    IOCTL_REQUEST(PTP_SYS_OFFSET2);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_DV_TIMINGS);
+    IOCTL_REQUEST(DMA_BUF_SET_NAME_A);
+    IOCTL_REQUEST(PTP_PIN_GETFUNC);
+    IOCTL_REQUEST(PTP_SYS_OFFSET_EXTENDED);
+    IOCTL_REQUEST(DFL_FPGA_PORT_UINT_SET_IRQ);
+    IOCTL_REQUEST(RTC_EPOCH_READ);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_SELECTION);
+    IOCTL_REQUEST(VIDIOC_QUERY_EXT_CTRL);
+    IOCTL_REQUEST(ATM_GETLECSADDR);
+    IOCTL_REQUEST(FSL_HV_IOCTL_PARTITION_STOP);
+    IOCTL_REQUEST(SONET_GETDIAG);
+    IOCTL_REQUEST(ATMMPC_DATA);
+    IOCTL_REQUEST(IPMICTL_UNREGISTER_FOR_CMD_CHANS);
+    IOCTL_REQUEST(HIDIOCGCOLLECTIONINDEX);
+    IOCTL_REQUEST(RPMSG_CREATE_EPT_IOCTL);
+    IOCTL_REQUEST(GPIOHANDLE_GET_LINE_VALUES_IOCTL);
+    IOCTL_REQUEST(UI_DEV_SETUP);
+    IOCTL_REQUEST(ISST_IF_IO_CMD);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_READ_REMOTE);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_HIST_CFG);
+    IOCTL_REQUEST(BLKGETNRZONES);
+    IOCTL_REQUEST(VIDIOC_G_MODULATOR);
+    IOCTL_REQUEST(VBG_IOCTL_WRITE_CORE_DUMP);
+    IOCTL_REQUEST(USBDEVFS_SETINTERFACE);
+    IOCTL_REQUEST(PPPIOCGCHAN);
+    IOCTL_REQUEST(EVIOCGVERSION);
+    IOCTL_REQUEST(VHOST_NET_SET_BACKEND);
+    IOCTL_REQUEST(USBDEVFS_REAPURBNDELAY);
+    IOCTL_REQUEST(RNDZAPENTCNT);
+    IOCTL_REQUEST(VIDIOC_G_PARM);
+    IOCTL_REQUEST(TUNGETDEVNETNS);
+    IOCTL_REQUEST(LIRC_SET_MEASURE_CARRIER_MODE);
+    IOCTL_REQUEST(VHOST_SET_VRING_ERR);
+#ifdef VDUSE_VQ_SETUP
+    IOCTL_REQUEST(VDUSE_VQ_SETUP);
+#endif
+    IOCTL_REQUEST(AUTOFS_IOC_SETTIMEOUT);
+    IOCTL_REQUEST(VIDIOC_S_FREQUENCY);
+    IOCTL_REQUEST(F2FS_IOC_SEC_TRIM_FILE);
+    IOCTL_REQUEST(FS_IOC_REMOVE_ENCRYPTION_KEY);
+    IOCTL_REQUEST(WDIOC_GETPRETIMEOUT);
+    IOCTL_REQUEST(USBDEVFS_DROP_PRIVILEGES);
+    IOCTL_REQUEST(BTRFS_IOC_SNAP_CREATE_V2);
+    IOCTL_REQUEST(VHOST_VSOCK_SET_RUNNING);
+    IOCTL_REQUEST(STP_SET_OPTIONS);
+    IOCTL_REQUEST(FBIO_RADEON_GET_MIRROR);
+    IOCTL_REQUEST(IVTVFB_IOC_DMA_FRAME);
+    IOCTL_REQUEST(IPMICTL_SEND_COMMAND);
+    IOCTL_REQUEST(VIDIOC_G_ENC_INDEX);
+    IOCTL_REQUEST(DFL_FPGA_FME_PORT_PR);
+    IOCTL_REQUEST(CHIOSVOLTAG);
+    IOCTL_REQUEST(ATM_SETESIF);
+    IOCTL_REQUEST(FW_CDEV_IOC_SEND_RESPONSE);
+    IOCTL_REQUEST(PMU_IOC_GET_MODEL);
+    IOCTL_REQUEST(JSIOCGBTNMAP);
+    IOCTL_REQUEST(USBDEVFS_HUB_PORTINFO);
+    IOCTL_REQUEST(VBG_IOCTL_INTERRUPT_ALL_WAIT_FOR_EVENTS);
+    IOCTL_REQUEST(FDCLRPRM);
+    IOCTL_REQUEST(BTRFS_IOC_SCRUB);
+    IOCTL_REQUEST(USBDEVFS_DISCONNECT);
+    IOCTL_REQUEST(TUNSETVNETBE);
+    IOCTL_REQUEST(ATMTCP_REMOVE);
+    IOCTL_REQUEST(VHOST_VDPA_GET_CONFIG);
+    IOCTL_REQUEST(PPPIOCGNPMODE);
+    IOCTL_REQUEST(FDGETDRVPRM);
+    IOCTL_REQUEST(TUNSETVNETLE);
+    IOCTL_REQUEST(PHN_SETREG);
+    IOCTL_REQUEST(PPPIOCDETACH);
+    IOCTL_REQUEST(MMTIMER_GETRES);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_ENUMSTD);
+    IOCTL_REQUEST(PPGETFLAGS);
+    IOCTL_REQUEST(VDUSE_DEV_GET_FEATURES);
+    IOCTL_REQUEST(CAPI_MANUFACTURER_CMD);
+    IOCTL_REQUEST(VIDIOC_G_TUNER);
+    IOCTL_REQUEST(DM_TABLE_STATUS);
+    IOCTL_REQUEST(DM_DEV_ARM_POLL);
+    IOCTL_REQUEST(NE_CREATE_VM);
+    IOCTL_REQUEST(MEDIA_IOC_ENUM_LINKS);
+    IOCTL_REQUEST(F2FS_IOC_PRECACHE_EXTENTS);
+    IOCTL_REQUEST(DFL_FPGA_PORT_DMA_MAP);
+    IOCTL_REQUEST(MGSL_IOCGXCTRL);
+    IOCTL_REQUEST(FW_CDEV_IOC_SEND_REQUEST);
+    IOCTL_REQUEST(SONYPI_IOCGBLUE);
+    IOCTL_REQUEST(F2FS_IOC_DECOMPRESS_FILE);
+    IOCTL_REQUEST(I2OHTML);
+    IOCTL_REQUEST(VFIO_GET_API_VERSION);
+    IOCTL_REQUEST(IDT77105_GETSTATZ);
+    IOCTL_REQUEST(I2OPARMSET);
+    IOCTL_REQUEST(TEE_IOC_CANCEL);
+    IOCTL_REQUEST(PTP_SYS_OFFSET_PRECISE2);
+    IOCTL_REQUEST(DFL_FPGA_PORT_RESET);
+    IOCTL_REQUEST(PPPIOCGASYNCMAP);
+    IOCTL_REQUEST(EVIOCGKEYCODE_V2);
+    IOCTL_REQUEST(DM_DEV_SET_GEOMETRY);
+    IOCTL_REQUEST(HIDIOCSUSAGE);
+    IOCTL_REQUEST(FW_CDEV_IOC_DEALLOCATE_ISO_RESOURCE_ONCE);
+    IOCTL_REQUEST(PTP_EXTTS_REQUEST);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_EVENT_CTL);
+    IOCTL_REQUEST(WDIOC_SETPRETIMEOUT);
+    IOCTL_REQUEST(VHOST_SCSI_CLEAR_ENDPOINT);
+    IOCTL_REQUEST(JSIOCGAXES);
+    IOCTL_REQUEST(HIDIOCSFLAG);
+    IOCTL_REQUEST(PTP_PEROUT_REQUEST2);
+    IOCTL_REQUEST(PPWDATA);
+    IOCTL_REQUEST(PTP_CLOCK_GETCAPS);
+    IOCTL_REQUEST(FDGETMAXERRS);
+    IOCTL_REQUEST(TUNSETQUEUE);
+    IOCTL_REQUEST(PTP_ENABLE_PPS);
+    IOCTL_REQUEST(SIOCSIFATMTCP);
+    IOCTL_REQUEST(CEC_ADAP_G_LOG_ADDRS);
+    IOCTL_REQUEST(ND_IOCTL_ARS_CAP);
+    IOCTL_REQUEST(NBD_SET_BLKSIZE);
+    IOCTL_REQUEST(NBD_SET_TIMEOUT);
+    IOCTL_REQUEST(VHOST_SCSI_GET_ABI_VERSION);
+    IOCTL_REQUEST(RIO_UNMAP_INBOUND);
+    IOCTL_REQUEST(ATM_QUERYLOOP);
+    IOCTL_REQUEST(DFL_FPGA_GET_API_VERSION);
+    IOCTL_REQUEST(USBDEVFS_WAIT_FOR_RESUME);
+    IOCTL_REQUEST(FBIO_CURSOR);
+    IOCTL_REQUEST(RNDCLEARPOOL);
+    IOCTL_REQUEST(VIDIOC_QUERYSTD);
+    IOCTL_REQUEST(DMA_BUF_IOCTL_SYNC);
+    IOCTL_REQUEST(SCIF_RECV);
+    IOCTL_REQUEST(PTP_PIN_GETFUNC2);
+    IOCTL_REQUEST(FW_CDEV_IOC_ALLOCATE);
+    IOCTL_REQUEST(CEC_ADAP_G_CAPS);
+    IOCTL_REQUEST(VIDIOC_G_FBUF);
+    IOCTL_REQUEST(PTP_ENABLE_PPS2);
+    IOCTL_REQUEST(PCITEST_CLEAR_IRQ);
+    IOCTL_REQUEST(IPMICTL_SET_GETS_EVENTS_CMD);
+    IOCTL_REQUEST(BTRFS_IOC_DEVICES_READY);
+    IOCTL_REQUEST(JSIOCGAXMAP);
+    IOCTL_REQUEST(FW_CDEV_IOC_GET_CYCLE_TIMER);
+    IOCTL_REQUEST(FW_CDEV_IOC_SET_ISO_CHANNELS);
+    IOCTL_REQUEST(RTC_WIE_OFF);
+    IOCTL_REQUEST(PPGETMODE);
+    IOCTL_REQUEST(VIDIOC_DBG_G_REGISTER);
+    IOCTL_REQUEST(PTP_SYS_OFFSET);
+    IOCTL_REQUEST(BTRFS_IOC_SPACE_INFO);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_ENUM_FRAME_SIZE);
+    IOCTL_REQUEST(ND_IOCTL_VENDOR);
+    IOCTL_REQUEST(SIOCSNETADDR);
+    IOCTL_REQUEST(SCIF_VREADFROM);
+    IOCTL_REQUEST(BTRFS_IOC_TRANS_START);
+    IOCTL_REQUEST(INOTIFY_IOC_SETNEXTWD);
+    IOCTL_REQUEST(SNAPSHOT_GET_IMAGE_SIZE);
+    IOCTL_REQUEST(TUNDETACHFILTER);
+    IOCTL_REQUEST(ND_IOCTL_CLEAR_ERROR);
+    IOCTL_REQUEST(IOC_PR_CLEAR);
+    IOCTL_REQUEST(SCIF_READFROM);
+    IOCTL_REQUEST(PPPIOCGDEBUG);
+    IOCTL_REQUEST(BLKGETZONESZ);
+    IOCTL_REQUEST(HIDIOCGUSAGES);
+    IOCTL_REQUEST(SONYPI_IOCGTEMP);
+    IOCTL_REQUEST(UI_SET_MSCBIT);
+    IOCTL_REQUEST(APM_IOC_SUSPEND);
+    IOCTL_REQUEST(BTRFS_IOC_TREE_SEARCH);
+    IOCTL_REQUEST(RTC_PLL_GET);
+    IOCTL_REQUEST(RIO_CM_EP_GET_LIST);
+    IOCTL_REQUEST(USBDEVFS_DISCSIGNAL);
+    IOCTL_REQUEST(OSIOCSNETADDR);
+    IOCTL_REQUEST(LIRC_GET_MIN_TIMEOUT);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_EVENT_SUMMARY_LEGACY);
+    IOCTL_REQUEST(DM_TARGET_MSG);
+    IOCTL_REQUEST(SONYPI_IOCGBAT1REM);
+    IOCTL_REQUEST(EVIOCSFF);
+    IOCTL_REQUEST(TUNSETGROUP);
+    IOCTL_REQUEST(EVIOCGKEYCODE);
+    IOCTL_REQUEST(KCOV_REMOTE_ENABLE);
+    IOCTL_REQUEST(ND_IOCTL_GET_CONFIG_SIZE);
+    IOCTL_REQUEST(FDEJECT);
+    IOCTL_REQUEST(TUNSETOFFLOAD);
+    IOCTL_REQUEST(PPPIOCCONNECT);
+    IOCTL_REQUEST(ATM_ADDADDR);
+    IOCTL_REQUEST(VDUSE_DEV_INJECT_CONFIG_IRQ);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_ASKUMOUNT);
+    IOCTL_REQUEST(VHOST_VDPA_GET_STATUS);
+    IOCTL_REQUEST(CCISS_PASSTHRU);
+    IOCTL_REQUEST(MGSL_IOCCLRMODCOUNT);
+    IOCTL_REQUEST(TEE_IOC_SUPPL_SEND);
+    IOCTL_REQUEST(ATMARPD_CTRL);
+    IOCTL_REQUEST(UI_ABS_SETUP);
+    IOCTL_REQUEST(UI_DEV_DESTROY);
+    IOCTL_REQUEST(BTRFS_IOC_QUOTA_CTL);
+    IOCTL_REQUEST(RTC_AIE_ON);
+    IOCTL_REQUEST(AUTOFS_IOC_EXPIRE);
+    IOCTL_REQUEST(PPPIOCSDEBUG);
+    IOCTL_REQUEST(GPIO_V2_LINE_SET_VALUES_IOCTL);
+    IOCTL_REQUEST(PPPIOCSMRU);
+    IOCTL_REQUEST(CCISS_DEREGDISK);
+    IOCTL_REQUEST(UI_DEV_CREATE);
+    IOCTL_REQUEST(FUSE_DEV_IOC_CLONE);
+    IOCTL_REQUEST(BTRFS_IOC_START_SYNC);
+    IOCTL_REQUEST(NILFS_IOCTL_DELETE_CHECKPOINT);
+    IOCTL_REQUEST(SNAPSHOT_AVAIL_SWAP_SIZE);
+    IOCTL_REQUEST(DM_TABLE_CLEAR);
+    IOCTL_REQUEST(CCISS_GETINTINFO);
+    IOCTL_REQUEST(PPPIOCSASYNCMAP);
+    IOCTL_REQUEST(I2OEVTGET);
+    IOCTL_REQUEST(NVME_IOCTL_RESET);
+    IOCTL_REQUEST(PPYIELD);
+    IOCTL_REQUEST(NVME_IOCTL_IO64_CMD);
+    IOCTL_REQUEST(TUNSETCARRIER);
+    IOCTL_REQUEST(DM_DEV_WAIT);
+    IOCTL_REQUEST(RTC_WIE_ON);
+#if 0 // needs `struct usbdevfs_ioctl32`
+    IOCTL_REQUEST(USBDEVFS_IOCTL32);
+#endif
+    IOCTL_REQUEST(MEDIA_IOC_DEVICE_INFO);
+    IOCTL_REQUEST(RIO_CM_CHAN_CREATE);
+    IOCTL_REQUEST(MGSL_IOCSPARAMS);
+    IOCTL_REQUEST(RTC_SET_TIME);
+    IOCTL_REQUEST(VHOST_RESET_OWNER);
+    IOCTL_REQUEST(IOC_OPAL_PSID_REVERT_TPR);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_OPENMOUNT);
+    IOCTL_REQUEST(UDF_GETEABLOCK);
+    IOCTL_REQUEST(VFIO_IOMMU_MAP_DMA);
+    IOCTL_REQUEST(VIDIOC_SUBSCRIBE_EVENT);
+    IOCTL_REQUEST(HIDIOCGFLAG);
+    IOCTL_REQUEST(HIDIOCGUCODE);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_AF_CFG);
+    IOCTL_REQUEST(DM_REMOVE_ALL);
+    IOCTL_REQUEST(ASPEED_LPC_CTRL_IOCTL_MAP);
+    IOCTL_REQUEST(CCISS_GETFIRMVER);
+    IOCTL_REQUEST(ND_IOCTL_ARS_START);
+    IOCTL_REQUEST(PPPIOCSMRRU);
+    IOCTL_REQUEST(CEC_ADAP_S_LOG_ADDRS);
+    IOCTL_REQUEST(RPROC_GET_SHUTDOWN_ON_RELEASE);
+    IOCTL_REQUEST(DMA_HEAP_IOCTL_ALLOC);
+    IOCTL_REQUEST(PPSETTIME);
+    IOCTL_REQUEST(RTC_ALM_READ);
+    IOCTL_REQUEST(VDUSE_SET_API_VERSION);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_WRITE_REMOTE);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_CROP);
+    IOCTL_REQUEST(USBDEVFS_CONNECT);
+    IOCTL_REQUEST(SYNC_IOC_FILE_INFO);
+    IOCTL_REQUEST(ATMARP_MKIP);
+    IOCTL_REQUEST(VFIO_IOMMU_SPAPR_TCE_GET_INFO);
+    IOCTL_REQUEST(CCISS_GETHEARTBEAT);
+    IOCTL_REQUEST(ATM_RSTADDR);
+    IOCTL_REQUEST(NBD_SET_SIZE);
+    IOCTL_REQUEST(UDF_GETVOLIDENT);
+    IOCTL_REQUEST(GPIO_V2_LINE_GET_VALUES_IOCTL);
+    IOCTL_REQUEST(MGSL_IOCSTXIDLE);
+    IOCTL_REQUEST(FSL_HV_IOCTL_SETPROP);
+    IOCTL_REQUEST(BTRFS_IOC_GET_DEV_STATS);
+    IOCTL_REQUEST(PPRSTATUS);
+    IOCTL_REQUEST(MGSL_IOCTXENABLE);
+    IOCTL_REQUEST(UDF_GETEASIZE);
+    IOCTL_REQUEST(NVME_IOCTL_ADMIN64_CMD);
+    IOCTL_REQUEST(VHOST_SET_OWNER);
+    IOCTL_REQUEST(RIO_ALLOC_DMA);
+    IOCTL_REQUEST(F2FS_IOC_ABORT_VOLATILE_WRITE);
+    IOCTL_REQUEST(RIO_CM_CHAN_ACCEPT);
+    IOCTL_REQUEST(I2OHRTGET);
+    IOCTL_REQUEST(ATM_SETCIRANGE);
+    IOCTL_REQUEST(HPET_IE_ON);
+    IOCTL_REQUEST(PERF_EVENT_IOC_ID);
+    IOCTL_REQUEST(TUNSETSNDBUF);
+    IOCTL_REQUEST(PTP_PIN_SETFUNC);
+    IOCTL_REQUEST(PPPIOCDISCONN);
+    IOCTL_REQUEST(VIDIOC_QUERYCTRL);
+    IOCTL_REQUEST(PPEXCL);
+    IOCTL_REQUEST(PCITEST_MSI);
+    IOCTL_REQUEST(FDWERRORCLR);
+    IOCTL_REQUEST(AUTOFS_IOC_FAIL);
+    IOCTL_REQUEST(USBDEVFS_IOCTL);
+    IOCTL_REQUEST(VIDIOC_S_STD);
+    IOCTL_REQUEST(F2FS_IOC_RESIZE_FS);
+    IOCTL_REQUEST(SONET_SETDIAG);
+    IOCTL_REQUEST(BTRFS_IOC_DEFRAG);
+    IOCTL_REQUEST(CCISS_GETDRIVVER);
+    IOCTL_REQUEST(IPMICTL_GET_TIMING_PARMS_CMD);
+    IOCTL_REQUEST(HPET_IRQFREQ);
+    IOCTL_REQUEST(ATM_GETESI);
+    IOCTL_REQUEST(CCISS_GETLUNINFO);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_ISMOUNTPOINT);
+    IOCTL_REQUEST(TEE_IOC_SHM_ALLOC);
+    IOCTL_REQUEST(PERF_EVENT_IOC_SET_BPF);
+    IOCTL_REQUEST(UDMABUF_CREATE_LIST);
+    IOCTL_REQUEST(VHOST_SET_LOG_BASE);
+    IOCTL_REQUEST(ZATM_GETPOOL);
+    IOCTL_REQUEST(BR2684_SETFILT);
+    IOCTL_REQUEST(RNDGETPOOL);
+    IOCTL_REQUEST(PPS_GETPARAMS);
+    IOCTL_REQUEST(IOC_PR_RESERVE);
+    IOCTL_REQUEST(VIDIOC_TRY_DECODER_CMD);
+    IOCTL_REQUEST(RIO_CM_CHAN_CLOSE);
+    IOCTL_REQUEST(VIDIOC_DV_TIMINGS_CAP);
+    IOCTL_REQUEST(IOCTL_MEI_CONNECT_CLIENT_VTAG);
+    IOCTL_REQUEST(PMU_IOC_GET_BACKLIGHT);
+    IOCTL_REQUEST(USBDEVFS_GET_CAPABILITIES);
+    IOCTL_REQUEST(SCIF_WRITETO);
+    IOCTL_REQUEST(UDF_RELOCATE_BLOCKS);
+    IOCTL_REQUEST(FSL_HV_IOCTL_PARTITION_RESTART);
+    IOCTL_REQUEST(CCISS_REGNEWD);
+    IOCTL_REQUEST(FAT_IOCTL_SET_ATTRIBUTES);
+    IOCTL_REQUEST(VIDIOC_CREATE_BUFS);
+    IOCTL_REQUEST(CAPI_GET_VERSION);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_EVENT_SUMMARY);
+    IOCTL_REQUEST(VFIO_EEH_PE_OP);
+    IOCTL_REQUEST(FW_CDEV_IOC_CREATE_ISO_CONTEXT);
+    IOCTL_REQUEST(F2FS_IOC_RELEASE_COMPRESS_BLOCKS);
+    IOCTL_REQUEST(NBD_SET_SIZE_BLOCKS);
+    IOCTL_REQUEST(IPMI_BMC_IOCTL_SET_SMS_ATN);
+    IOCTL_REQUEST(ASPEED_P2A_CTRL_IOCTL_GET_MEMORY_CONFIG);
+    IOCTL_REQUEST(VIDIOC_S_AUDOUT);
+    IOCTL_REQUEST(VIDIOC_S_FMT);
+    IOCTL_REQUEST(PPPIOCATTACH);
+    IOCTL_REQUEST(VHOST_GET_VRING_BUSYLOOP_TIMEOUT);
+    IOCTL_REQUEST(FS_IOC_MEASURE_VERITY);
+    IOCTL_REQUEST(CCISS_BIG_PASSTHRU);
+    IOCTL_REQUEST(IPMICTL_SET_MY_LUN_CMD);
+    IOCTL_REQUEST(PCITEST_LEGACY_IRQ);
+    IOCTL_REQUEST(USBDEVFS_SUBMITURB);
+    IOCTL_REQUEST(AUTOFS_IOC_READY);
+    IOCTL_REQUEST(BTRFS_IOC_SEND);
+    IOCTL_REQUEST(VIDIOC_G_EXT_CTRLS);
+    IOCTL_REQUEST(JSIOCSBTNMAP);
+    IOCTL_REQUEST(PPPIOCSFLAGS);
+    IOCTL_REQUEST(NVRAM_INIT);
+    IOCTL_REQUEST(RFKILL_IOCTL_NOINPUT);
+    IOCTL_REQUEST(BTRFS_IOC_BALANCE);
+    IOCTL_REQUEST(FS_IOC_GETFSMAP);
+    IOCTL_REQUEST(IPMICTL_GET_MY_CHANNEL_LUN_CMD);
+    IOCTL_REQUEST(STP_POLICY_ID_GET);
+    IOCTL_REQUEST(PPSETFLAGS);
+    IOCTL_REQUEST(CEC_ADAP_S_PHYS_ADDR);
+    IOCTL_REQUEST(ATMTCP_CREATE);
+    IOCTL_REQUEST(IPMI_BMC_IOCTL_FORCE_ABORT);
+    IOCTL_REQUEST(PPPIOCGXASYNCMAP);
+    IOCTL_REQUEST(VHOST_SET_VRING_CALL);
+    IOCTL_REQUEST(LIRC_GET_FEATURES);
+    IOCTL_REQUEST(GSMIOC_DISABLE_NET);
+    IOCTL_REQUEST(AUTOFS_IOC_CATATONIC);
+    IOCTL_REQUEST(NBD_DO_IT);
+    IOCTL_REQUEST(LIRC_SET_REC_CARRIER_RANGE);
+    IOCTL_REQUEST(IPMICTL_GET_MY_CHANNEL_ADDRESS_CMD);
+    IOCTL_REQUEST(EVIOCSCLOCKID);
+    IOCTL_REQUEST(USBDEVFS_FREE_STREAMS);
+    IOCTL_REQUEST(FSI_SCOM_RESET);
+    IOCTL_REQUEST(PMU_IOC_GRAB_BACKLIGHT);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_FMT);
+    IOCTL_REQUEST(FDDEFPRM);
+    IOCTL_REQUEST(TEE_IOC_INVOKE);
+    IOCTL_REQUEST(USBDEVFS_BULK);
+    IOCTL_REQUEST(SCIF_VWRITETO);
+    IOCTL_REQUEST(SONYPI_IOCSBRT);
+    IOCTL_REQUEST(BTRFS_IOC_FILE_EXTENT_SAME);
+    IOCTL_REQUEST(RTC_PIE_ON);
+    IOCTL_REQUEST(BTRFS_IOC_SCAN_DEV);
+    IOCTL_REQUEST(PPPIOCXFERUNIT);
+    IOCTL_REQUEST(WDIOC_GETTIMEOUT);
+    IOCTL_REQUEST(BTRFS_IOC_SET_RECEIVED_SUBVOL);
+    IOCTL_REQUEST(DFL_FPGA_PORT_ERR_SET_IRQ);
+    IOCTL_REQUEST(FBIO_WAITFORVSYNC);
+    IOCTL_REQUEST(RTC_PIE_OFF);
+    IOCTL_REQUEST(EVIOCGRAB);
+    IOCTL_REQUEST(PMU_IOC_SET_BACKLIGHT);
+    IOCTL_REQUEST(EVIOCGREP);
+    IOCTL_REQUEST(PERF_EVENT_IOC_MODIFY_ATTRIBUTES);
+    IOCTL_REQUEST(UFFDIO_CONTINUE);
+    IOCTL_REQUEST(VDUSE_GET_API_VERSION);
+    IOCTL_REQUEST(RTC_RD_TIME);
+    IOCTL_REQUEST(FDMSGOFF);
+    IOCTL_REQUEST(IPMICTL_REGISTER_FOR_CMD_CHANS);
+    IOCTL_REQUEST(CAPI_GET_ERRCODE);
+    IOCTL_REQUEST(PCITEST_SET_IRQTYPE);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_EDID);
+    IOCTL_REQUEST(MATROXFB_SET_OUTPUT_MODE);
+    IOCTL_REQUEST(RIO_DEV_ADD);
+    IOCTL_REQUEST(VIDIOC_ENUM_FREQ_BANDS);
+    IOCTL_REQUEST(FBIO_RADEON_SET_MIRROR);
+    IOCTL_REQUEST(PCITEST_GET_IRQTYPE);
+#if 0 // needs `struct usbdevfs_urb32`
+    IOCTL_REQUEST(USBDEVFS_SUBMITURB32);
+#endif
+    IOCTL_REQUEST(JSIOCGVERSION);
+    IOCTL_REQUEST(SONYPI_IOCSBLUE);
+    IOCTL_REQUEST(SNAPSHOT_PREF_IMAGE_SIZE);
+    IOCTL_REQUEST(F2FS_IOC_GET_FEATURES);
+    IOCTL_REQUEST(SCIF_REG);
+    IOCTL_REQUEST(NILFS_IOCTL_CLEAN_SEGMENTS);
+    IOCTL_REQUEST(FW_CDEV_IOC_INITIATE_BUS_RESET);
+    IOCTL_REQUEST(RIO_WAIT_FOR_ASYNC);
+    IOCTL_REQUEST(VHOST_SET_VRING_NUM);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_PROTOVER);
+    IOCTL_REQUEST(RIO_FREE_DMA);
+#if 0 // needs `struct usbdevfs_bulktransfer32`
+    IOCTL_REQUEST(USBDEVFS_BULK32);
+#endif
+    IOCTL_REQUEST(MGSL_IOCRXENABLE);
+    IOCTL_REQUEST(IOCTL_VM_SOCKETS_GET_LOCAL_CID);
+    IOCTL_REQUEST(IPMICTL_SET_TIMING_PARMS_CMD);
+    IOCTL_REQUEST(PPPIOCGL2TPSTATS);
+    IOCTL_REQUEST(PERF_EVENT_IOC_PERIOD);
+    IOCTL_REQUEST(PTP_PIN_SETFUNC2);
+    IOCTL_REQUEST(CHIOEXCHANGE);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_SUINFO);
+    IOCTL_REQUEST(CEC_DQEVENT);
+    IOCTL_REQUEST(UI_SET_SWBIT);
+    IOCTL_REQUEST(VHOST_VDPA_SET_CONFIG);
+    IOCTL_REQUEST(TUNSETIFF);
+    IOCTL_REQUEST(CHIOPOSITION);
+    IOCTL_REQUEST(IPMICTL_SET_MAINTENANCE_MODE_CMD);
+    IOCTL_REQUEST(BTRFS_IOC_DEFAULT_SUBVOL);
+    IOCTL_REQUEST(RIO_UNMAP_OUTBOUND);
+    IOCTL_REQUEST(CAPI_CLR_FLAGS);
+    IOCTL_REQUEST(FW_CDEV_IOC_ALLOCATE_ISO_RESOURCE_ONCE);
+    IOCTL_REQUEST(MATROXFB_GET_OUTPUT_CONNECTION);
+    IOCTL_REQUEST(EVIOCSMASK);
+    IOCTL_REQUEST(BTRFS_IOC_FORGET_DEV);
+#ifdef CXL_MEM_QUERY_COMMANDS
+    IOCTL_REQUEST(CXL_MEM_QUERY_COMMANDS);
+#endif
+    IOCTL_REQUEST(CEC_S_MODE);
+    IOCTL_REQUEST(MGSL_IOCSIF);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_PFF_TO_PORT);
+    IOCTL_REQUEST(PPSETMODE);
+    IOCTL_REQUEST(VFIO_DEVICE_SET_IRQS);
+    IOCTL_REQUEST(VIDIOC_PREPARE_BUF);
+    IOCTL_REQUEST(CEC_ADAP_G_CONNECTOR_INFO);
+    IOCTL_REQUEST(IOC_OPAL_WRITE_SHADOW_MBR);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_ENUM_FRAME_INTERVAL);
+    IOCTL_REQUEST(UDMABUF_CREATE);
+    IOCTL_REQUEST(SONET_CLRDIAG);
+    IOCTL_REQUEST(PHN_SET_REG);
+    IOCTL_REQUEST(RNDADDTOENTCNT);
+    IOCTL_REQUEST(VBG_IOCTL_CHECK_BALLOON);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_STAT_REQ);
+    IOCTL_REQUEST(PPS_FETCH);
+    IOCTL_REQUEST(RTC_AIE_OFF);
+    IOCTL_REQUEST(VFIO_GROUP_SET_CONTAINER);
+    IOCTL_REQUEST(FW_CDEV_IOC_RECEIVE_PHY_PACKETS);
+    IOCTL_REQUEST(VFIO_IOMMU_SPAPR_TCE_REMOVE);
+    IOCTL_REQUEST(VFIO_IOMMU_GET_INFO);
+#if 0 // needs `struct omap3isp_stat_data_time32`
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_STAT_REQ_TIME32);
+#endif
+    IOCTL_REQUEST(DM_DEV_SUSPEND);
+    IOCTL_REQUEST(F2FS_IOC_GET_COMPRESS_OPTION);
+    IOCTL_REQUEST(FW_CDEV_IOC_STOP_ISO);
+    IOCTL_REQUEST(GPIO_V2_GET_LINEINFO_IOCTL);
+    IOCTL_REQUEST(ATMMPC_CTRL);
+    IOCTL_REQUEST(PPPIOCSXASYNCMAP);
+    IOCTL_REQUEST(CHIOGSTATUS);
+    IOCTL_REQUEST(FW_CDEV_IOC_ALLOCATE_ISO_RESOURCE);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_PORT_IDX_GET);
+    IOCTL_REQUEST(CAPI_SET_FLAGS);
+    IOCTL_REQUEST(VFIO_GROUP_GET_DEVICE_FD);
+    IOCTL_REQUEST(VHOST_SET_MEM_TABLE);
+    IOCTL_REQUEST(MATROXFB_SET_OUTPUT_CONNECTION);
+    IOCTL_REQUEST(DFL_FPGA_PORT_GET_REGION_INFO);
+    IOCTL_REQUEST(VHOST_GET_FEATURES);
+    IOCTL_REQUEST(LIRC_GET_REC_RESOLUTION);
+    IOCTL_REQUEST(PACKET_CTRL_CMD);
+    IOCTL_REQUEST(LIRC_SET_TRANSMITTER_MASK);
+    IOCTL_REQUEST(BTRFS_IOC_ADD_DEV);
+    IOCTL_REQUEST(JSIOCGCORR);
+    IOCTL_REQUEST(VIDIOC_G_FMT);
+    IOCTL_REQUEST(RTC_EPOCH_SET);
+    IOCTL_REQUEST(CAPI_GET_PROFILE);
+    IOCTL_REQUEST(ATM_GETLOOP);
+    IOCTL_REQUEST(SCIF_LISTEN);
+    IOCTL_REQUEST(NBD_CLEAR_QUE);
+    IOCTL_REQUEST(F2FS_IOC_MOVE_RANGE);
+    IOCTL_REQUEST(LIRC_GET_LENGTH);
+    IOCTL_REQUEST(I8K_SET_FAN);
+    IOCTL_REQUEST(FDSETMAXERRS);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_QUERYCAP);
+    IOCTL_REQUEST(SNAPSHOT_SET_SWAP_AREA);
+    IOCTL_REQUEST(LIRC_GET_REC_TIMEOUT);
+    IOCTL_REQUEST(EVIOCRMFF);
+    IOCTL_REQUEST(GPIO_GET_LINEEVENT_IOCTL);
+    IOCTL_REQUEST(PPRDATA);
+    IOCTL_REQUEST(RIO_MPORT_GET_PROPERTIES);
+    IOCTL_REQUEST(TUNSETVNETHDRSZ);
+    IOCTL_REQUEST(GPIO_GET_LINEINFO_IOCTL);
+    IOCTL_REQUEST(GSMIOC_GETCONF);
+    IOCTL_REQUEST(LIRC_GET_SEND_MODE);
+    IOCTL_REQUEST(PPPIOCSACTIVE);
+    IOCTL_REQUEST(SIOCGSTAMPNS_NEW);
+    IOCTL_REQUEST(IPMICTL_RECEIVE_MSG);
+    IOCTL_REQUEST(LIRC_SET_SEND_DUTY_CYCLE);
+    IOCTL_REQUEST(UI_END_FF_ERASE);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_FLASH_PART_INFO);
+    IOCTL_REQUEST(FW_CDEV_IOC_SEND_PHY_PACKET);
+    IOCTL_REQUEST(NBD_SET_FLAGS);
+    IOCTL_REQUEST(SIOCGNETADDR);
+    IOCTL_REQUEST(VFIO_DEVICE_GET_REGION_INFO);
+    IOCTL_REQUEST(REISERFS_IOC_UNPACK);
+    IOCTL_REQUEST(FW_CDEV_IOC_REMOVE_DESCRIPTOR);
+    IOCTL_REQUEST(RIO_SET_EVENT_MASK);
+    IOCTL_REQUEST(SNAPSHOT_ALLOC_SWAP_PAGE);
+    IOCTL_REQUEST(VDUSE_VQ_INJECT_IRQ);
+    IOCTL_REQUEST(I2OPASSTHRU);
+    IOCTL_REQUEST(IOC_OPAL_SET_PW);
+    IOCTL_REQUEST(FSI_SCOM_READ);
+    IOCTL_REQUEST(VHOST_VDPA_GET_DEVICE_ID);
+    IOCTL_REQUEST(VIDIOC_QBUF);
+    IOCTL_REQUEST(VIDIOC_S_TUNER);
+    IOCTL_REQUEST(TUNGETVNETHDRSZ);
+    IOCTL_REQUEST(CAPI_NCCI_GETUNIT);
+    IOCTL_REQUEST(DFL_FPGA_PORT_UINT_GET_IRQ_NUM);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_STAT_EN);
+    IOCTL_REQUEST(GPIO_V2_LINE_SET_CONFIG_IOCTL);
+    IOCTL_REQUEST(TEE_IOC_VERSION);
+    IOCTL_REQUEST(VIDIOC_LOG_STATUS);
+    IOCTL_REQUEST(IPMICTL_SEND_COMMAND_SETTIME);
+    IOCTL_REQUEST(VHOST_SET_LOG_FD);
+    IOCTL_REQUEST(SCIF_SEND);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_FMT);
+    IOCTL_REQUEST(NS_ADJBUFLEV);
+    IOCTL_REQUEST(VIDIOC_DBG_S_REGISTER);
+    IOCTL_REQUEST(NILFS_IOCTL_RESIZE);
+    IOCTL_REQUEST(PHN_GETREG);
+    IOCTL_REQUEST(I2OSWDL);
+    IOCTL_REQUEST(VBG_IOCTL_VMMDEV_REQUEST_BIG);
+    IOCTL_REQUEST(OSIOCGNETADDR);
+    IOCTL_REQUEST(JSIOCGBUTTONS);
+    IOCTL_REQUEST(VFIO_IOMMU_ENABLE);
+    IOCTL_REQUEST(DM_DEV_RENAME);
+    IOCTL_REQUEST(MEDIA_IOC_SETUP_LINK);
+    IOCTL_REQUEST(VIDIOC_ENUMOUTPUT);
+    IOCTL_REQUEST(STP_POLICY_ID_SET);
+    IOCTL_REQUEST(VHOST_VDPA_SET_CONFIG_CALL);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_CROP);
+    IOCTL_REQUEST(VIDIOC_S_CROP);
+    IOCTL_REQUEST(WDIOC_GETTEMP);
+    IOCTL_REQUEST(IOC_OPAL_ADD_USR_TO_LR);
+    IOCTL_REQUEST(UI_SET_LEDBIT);
+    IOCTL_REQUEST(NBD_SET_SOCK);
+    IOCTL_REQUEST(BTRFS_IOC_SNAP_DESTROY_V2);
+    IOCTL_REQUEST(HIDIOCGCOLLECTIONINFO);
+    IOCTL_REQUEST(I2OSWUL);
+    IOCTL_REQUEST(IOCTL_MEI_NOTIFY_GET);
+    IOCTL_REQUEST(FDFMTTRK);
+    IOCTL_REQUEST(MMTIMER_GETBITS);
+    IOCTL_REQUEST(VIDIOC_ENUMSTD);
+    IOCTL_REQUEST(VHOST_GET_VRING_BASE);
+    IOCTL_REQUEST(VFIO_DEVICE_IOEVENTFD);
+    IOCTL_REQUEST(ATMARP_SETENTRY);
+    IOCTL_REQUEST(CCISS_REVALIDVOLS);
+    IOCTL_REQUEST(MGSL_IOCLOOPTXDONE);
+    IOCTL_REQUEST(RTC_VL_READ);
+    IOCTL_REQUEST(ND_IOCTL_ARS_STATUS);
+    IOCTL_REQUEST(RIO_DEV_DEL);
+    IOCTL_REQUEST(VBG_IOCTL_ACQUIRE_GUEST_CAPABILITIES);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_DV_TIMINGS_CAP);
+    IOCTL_REQUEST(SONYPI_IOCSFAN);
+    IOCTL_REQUEST(SPIOCSTYPE);
+    IOCTL_REQUEST(IPMICTL_REGISTER_FOR_CMD);
+    IOCTL_REQUEST(I8K_GET_FAN);
+    IOCTL_REQUEST(TUNGETVNETBE);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_FAIL);
+    IOCTL_REQUEST(UI_END_FF_UPLOAD);
+    IOCTL_REQUEST(TOSH_SMM);
+    IOCTL_REQUEST(SONYPI_IOCGBAT2REM);
+    IOCTL_REQUEST(F2FS_IOC_GET_COMPRESS_BLOCKS);
+    IOCTL_REQUEST(PPPIOCSNPMODE);
+    IOCTL_REQUEST(USBDEVFS_CONTROL);
+    IOCTL_REQUEST(HIDIOCGUSAGE);
+    IOCTL_REQUEST(TUNSETTXFILTER);
+    IOCTL_REQUEST(TUNGETVNETLE);
+    IOCTL_REQUEST(VIDIOC_ENUM_DV_TIMINGS);
+    IOCTL_REQUEST(BTRFS_IOC_INO_PATHS);
+    IOCTL_REQUEST(MGSL_IOCGXSYNC);
+    IOCTL_REQUEST(HIDIOCGFIELDINFO);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_STD);
+    IOCTL_REQUEST(I2OVALIDATE);
+    IOCTL_REQUEST(VIDIOC_TRY_ENCODER_CMD);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_CPINFO);
+    IOCTL_REQUEST(VIDIOC_G_FREQUENCY);
+    IOCTL_REQUEST(VFAT_IOCTL_READDIR_SHORT);
+    IOCTL_REQUEST(ND_IOCTL_GET_CONFIG_DATA);
+    IOCTL_REQUEST(F2FS_IOC_RESERVE_COMPRESS_BLOCKS);
+    IOCTL_REQUEST(FDGETDRVSTAT);
+    IOCTL_REQUEST(SYNC_IOC_MERGE);
+    IOCTL_REQUEST(VIDIOC_S_DV_TIMINGS);
+    IOCTL_REQUEST(PPPIOCBRIDGECHAN);
+    IOCTL_REQUEST(LIRC_SET_SEND_MODE);
+    IOCTL_REQUEST(RIO_ENABLE_PORTWRITE_RANGE);
+    IOCTL_REQUEST(ATM_GETTYPE);
+    IOCTL_REQUEST(PHN_GETREGS);
+    IOCTL_REQUEST(FDSETEMSGTRESH);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_VINFO);
+    IOCTL_REQUEST(MGSL_IOCWAITEVENT);
+    IOCTL_REQUEST(CAPI_INSTALLED);
+    IOCTL_REQUEST(EVIOCGMASK);
+    IOCTL_REQUEST(BTRFS_IOC_SUBVOL_GETFLAGS);
+    IOCTL_REQUEST(FSL_HV_IOCTL_PARTITION_GET_STATUS);
+    IOCTL_REQUEST(MEDIA_IOC_ENUM_ENTITIES);
+    IOCTL_REQUEST(GSMIOC_GETFIRST);
+    IOCTL_REQUEST(FW_CDEV_IOC_FLUSH_ISO);
+    IOCTL_REQUEST(VIDIOC_DBG_G_CHIP_INFO);
+    IOCTL_REQUEST(F2FS_IOC_RELEASE_VOLATILE_WRITE);
+    IOCTL_REQUEST(CAPI_GET_SERIAL);
+    IOCTL_REQUEST(FDSETDRVPRM);
+    IOCTL_REQUEST(IOC_OPAL_SAVE);
+    IOCTL_REQUEST(VIDIOC_G_DV_TIMINGS);
+    IOCTL_REQUEST(TUNSETIFINDEX);
+    IOCTL_REQUEST(CCISS_SETINTINFO);
+    IOCTL_REQUEST(CM_IOSDBGLVL);
+    IOCTL_REQUEST(RTC_VL_CLR);
+    IOCTL_REQUEST(VIDIOC_REQBUFS);
+    IOCTL_REQUEST(USBDEVFS_REAPURBNDELAY32);
+    IOCTL_REQUEST(TEE_IOC_SHM_REGISTER);
+    IOCTL_REQUEST(USBDEVFS_SETCONFIGURATION);
+    IOCTL_REQUEST(CCISS_GETNODENAME);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_FRAME_INTERVAL);
+    IOCTL_REQUEST(VIDIOC_ENUM_FRAMESIZES);
+    IOCTL_REQUEST(VFIO_DEVICE_PCI_HOT_RESET);
+    IOCTL_REQUEST(FW_CDEV_IOC_SEND_BROADCAST_REQUEST);
+    IOCTL_REQUEST(LPSETTIMEOUT_NEW);
+    IOCTL_REQUEST(RIO_CM_MPORT_GET_LIST);
+    IOCTL_REQUEST(FW_CDEV_IOC_QUEUE_ISO);
+    IOCTL_REQUEST(FDRAWCMD);
+    IOCTL_REQUEST(SCIF_UNREG);
+#if 0 // needs `struct serial_iso7816`
+    IOCTL_REQUEST(TIOCGISO7816);
+#endif
+    IOCTL_REQUEST(PPPIOCGIDLE64);
+    IOCTL_REQUEST(USBDEVFS_RELEASEINTERFACE);
+    IOCTL_REQUEST(VIDIOC_CROPCAP);
+    IOCTL_REQUEST(DFL_FPGA_PORT_GET_INFO);
+    IOCTL_REQUEST(PHN_SET_REGS);
+    IOCTL_REQUEST(ATMLEC_DATA);
+    IOCTL_REQUEST(PPPOEIOCDFWD);
+    IOCTL_REQUEST(VIDIOC_S_SELECTION);
+    IOCTL_REQUEST(SNAPSHOT_FREE_SWAP_PAGES);
+    IOCTL_REQUEST(BTRFS_IOC_LOGICAL_INO);
+    IOCTL_REQUEST(VIDIOC_S_CTRL);
+    IOCTL_REQUEST(ZATM_SETPOOL);
+    IOCTL_REQUEST(MTIOCPOS);
+    IOCTL_REQUEST(PMU_IOC_SLEEP);
+#if 0 // needs `struct usbdevfs_ctrltransfer32`
+    IOCTL_REQUEST(USBDEVFS_CONTROL32);
+#endif
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_PROTOSUBVER);
+    IOCTL_REQUEST(VBG_IOCTL_CHANGE_FILTER_MASK);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_SUSTAT);
+    IOCTL_REQUEST(VIDIOC_QUERYCAP);
+    IOCTL_REQUEST(HPET_INFO);
+    IOCTL_REQUEST(VIDIOC_AM437X_CCDC_CFG);
+    IOCTL_REQUEST(DM_LIST_DEVICES);
+    IOCTL_REQUEST(TUNSETOWNER);
+    IOCTL_REQUEST(VBG_IOCTL_CHANGE_GUEST_CAPABILITIES);
+    IOCTL_REQUEST(RNDADDENTROPY);
+    IOCTL_REQUEST(USBDEVFS_RESET);
+    IOCTL_REQUEST(BTRFS_IOC_SUBVOL_CREATE);
+    IOCTL_REQUEST(USBDEVFS_FORBID_SUSPEND);
+    IOCTL_REQUEST(FDGETDRVTYP);
+    IOCTL_REQUEST(PPWCONTROL);
+    IOCTL_REQUEST(VIDIOC_ENUM_FRAMEINTERVALS);
+    IOCTL_REQUEST(KCOV_DISABLE);
+    IOCTL_REQUEST(IOC_OPAL_ACTIVATE_LSP);
+    IOCTL_REQUEST(VHOST_VDPA_GET_IOVA_RANGE);
+    IOCTL_REQUEST(PPPIOCSPASS);
+    IOCTL_REQUEST(RIO_CM_CHAN_CONNECT);
+    IOCTL_REQUEST(I2OSWDEL);
+    IOCTL_REQUEST(FS_IOC_SET_ENCRYPTION_POLICY);
+    IOCTL_REQUEST(IOC_OPAL_MBR_DONE);
+    IOCTL_REQUEST(PPPIOCSMAXCID);
+    IOCTL_REQUEST(PPSETPHASE);
+    IOCTL_REQUEST(VHOST_VDPA_SET_VRING_ENABLE);
+    IOCTL_REQUEST(USBDEVFS_GET_SPEED);
+    IOCTL_REQUEST(SONET_GETFRAMING);
+    IOCTL_REQUEST(VIDIOC_QUERYBUF);
+    IOCTL_REQUEST(VIDIOC_S_EDID);
+    IOCTL_REQUEST(BTRFS_IOC_QGROUP_ASSIGN);
+    IOCTL_REQUEST(PPS_GETCAP);
+    IOCTL_REQUEST(SNAPSHOT_PLATFORM_SUPPORT);
+    IOCTL_REQUEST(LIRC_SET_REC_TIMEOUT_REPORTS);
+    IOCTL_REQUEST(SCIF_GET_NODEIDS);
+    IOCTL_REQUEST(NBD_DISCONNECT);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_FRAME_INTERVAL);
+    IOCTL_REQUEST(VFIO_IOMMU_DISABLE);
+    IOCTL_REQUEST(SNAPSHOT_CREATE_IMAGE);
+    IOCTL_REQUEST(SNAPSHOT_POWER_OFF);
+    IOCTL_REQUEST(APM_IOC_STANDBY);
+    IOCTL_REQUEST(PPPIOCGUNIT);
+    IOCTL_REQUEST(AUTOFS_IOC_EXPIRE_MULTI);
+    IOCTL_REQUEST(SCIF_BIND);
+    IOCTL_REQUEST(IOC_WATCH_QUEUE_SET_SIZE);
+    IOCTL_REQUEST(NILFS_IOCTL_CHANGE_CPMODE);
+    IOCTL_REQUEST(IOC_OPAL_LOCK_UNLOCK);
+    IOCTL_REQUEST(F2FS_IOC_SET_PIN_FILE);
+    IOCTL_REQUEST(PPPIOCGRASYNCMAP);
+    IOCTL_REQUEST(MMTIMER_MMAPAVAIL);
+    IOCTL_REQUEST(I2OPASSTHRU32);
+    IOCTL_REQUEST(DFL_FPGA_FME_PORT_RELEASE);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_QUERY_DV_TIMINGS);
+    IOCTL_REQUEST(UI_SET_SNDBIT);
+    IOCTL_REQUEST(VIDIOC_G_AUDOUT);
+    IOCTL_REQUEST(RTC_PLL_SET);
+    IOCTL_REQUEST(VIDIOC_ENUMAUDIO);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_TIMEOUT);
+    IOCTL_REQUEST(VBG_IOCTL_DRIVER_VERSION_INFO);
+    IOCTL_REQUEST(VHOST_SCSI_GET_EVENTS_MISSED);
+    IOCTL_REQUEST(VHOST_SET_VRING_ADDR);
+    IOCTL_REQUEST(VDUSE_CREATE_DEV);
+    IOCTL_REQUEST(FDFLUSH);
+    IOCTL_REQUEST(VBG_IOCTL_WAIT_FOR_EVENTS);
+    IOCTL_REQUEST(DFL_FPGA_FME_ERR_SET_IRQ);
+    IOCTL_REQUEST(F2FS_IOC_GET_PIN_FILE);
+    IOCTL_REQUEST(SCIF_CONNECT);
+    IOCTL_REQUEST(BLKREPORTZONE);
+    IOCTL_REQUEST(AUTOFS_IOC_ASKUMOUNT);
+    IOCTL_REQUEST(ATM_ADDPARTY);
+    IOCTL_REQUEST(FDSETPRM);
+    IOCTL_REQUEST(ATM_GETSTATZ);
+    IOCTL_REQUEST(ISST_IF_MSR_COMMAND);
+    IOCTL_REQUEST(BTRFS_IOC_GET_SUBVOL_INFO);
+    IOCTL_REQUEST(VIDIOC_UNSUBSCRIBE_EVENT);
+    IOCTL_REQUEST(SEV_ISSUE_CMD);
+    IOCTL_REQUEST(GPIOHANDLE_SET_LINE_VALUES_IOCTL);
+    IOCTL_REQUEST(PCITEST_COPY);
+    IOCTL_REQUEST(IPMICTL_GET_MY_ADDRESS_CMD);
+    IOCTL_REQUEST(CHIOGPICKER);
+    IOCTL_REQUEST(CAPI_NCCI_OPENCOUNT);
+    IOCTL_REQUEST(CXL_MEM_SEND_COMMAND);
+    IOCTL_REQUEST(PERF_EVENT_IOC_SET_FILTER);
+    IOCTL_REQUEST(IOC_OPAL_REVERT_TPR);
+    IOCTL_REQUEST(CHIOGVPARAMS);
+    IOCTL_REQUEST(PTP_PEROUT_REQUEST);
+    IOCTL_REQUEST(FSI_SCOM_CHECK);
+    IOCTL_REQUEST(RTC_IRQP_READ);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_READ_LOCAL);
+    IOCTL_REQUEST(HIDIOCGRDESCSIZE);
+    IOCTL_REQUEST(UI_GET_VERSION);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_CPSTAT);
+    IOCTL_REQUEST(CCISS_GETBUSTYPES);
+    IOCTL_REQUEST(VFIO_IOMMU_SPAPR_TCE_CREATE);
+    IOCTL_REQUEST(VIDIOC_EXPBUF);
+    IOCTL_REQUEST(UI_SET_RELBIT);
+    IOCTL_REQUEST(VFIO_SET_IOMMU);
+    IOCTL_REQUEST(VIDIOC_S_MODULATOR);
+    IOCTL_REQUEST(TUNGETFILTER);
+    IOCTL_REQUEST(MEYEIOC_SYNC);
+    IOCTL_REQUEST(CCISS_SETNODENAME);
+    IOCTL_REQUEST(FBIO_GETCONTROL2);
+    IOCTL_REQUEST(TUNSETDEBUG);
+    IOCTL_REQUEST(DM_DEV_REMOVE);
+    IOCTL_REQUEST(HIDIOCSUSAGES);
+    IOCTL_REQUEST(FS_IOC_ADD_ENCRYPTION_KEY);
+    IOCTL_REQUEST(FBIOGET_VBLANK);
+    IOCTL_REQUEST(ATM_GETSTAT);
+    IOCTL_REQUEST(VIDIOC_G_JPEGCOMP);
+    IOCTL_REQUEST(TUNATTACHFILTER);
+    IOCTL_REQUEST(UI_SET_ABSBIT);
+    IOCTL_REQUEST(DFL_FPGA_PORT_ERR_GET_IRQ_NUM);
+    IOCTL_REQUEST(USBDEVFS_REAPURB32);
+    IOCTL_REQUEST(BTRFS_IOC_TRANS_END);
+    IOCTL_REQUEST(CAPI_REGISTER);
+    IOCTL_REQUEST(F2FS_IOC_COMPRESS_FILE);
+    IOCTL_REQUEST(USBDEVFS_DISCARDURB);
+    IOCTL_REQUEST(HE_GET_REG);
+    IOCTL_REQUEST(ATM_SETLOOP);
+    IOCTL_REQUEST(ATMSIGD_CTRL);
+    IOCTL_REQUEST(CIOC_KERNEL_VERSION);
+    IOCTL_REQUEST(BTRFS_IOC_CLONE_RANGE);
+    IOCTL_REQUEST(SNAPSHOT_UNFREEZE);
+    IOCTL_REQUEST(F2FS_IOC_START_VOLATILE_WRITE);
+    IOCTL_REQUEST(PMU_IOC_HAS_ADB);
+    IOCTL_REQUEST(I2OGETIOPS);
+    IOCTL_REQUEST(VIDIOC_S_FBUF);
+    IOCTL_REQUEST(PPRCONTROL);
+    IOCTL_REQUEST(CHIOSPICKER);
+    IOCTL_REQUEST(VFIO_IOMMU_SPAPR_REGISTER_MEMORY);
+    IOCTL_REQUEST(TUNGETSNDBUF);
+    IOCTL_REQUEST(GSMIOC_SETCONF);
+    IOCTL_REQUEST(IOC_PR_PREEMPT);
+    IOCTL_REQUEST(KCOV_INIT_TRACE);
+    IOCTL_REQUEST(SONYPI_IOCGBAT1CAP);
+    IOCTL_REQUEST(SWITCHTEC_IOCTL_FLASH_INFO);
+    IOCTL_REQUEST(MTIOCTOP);
+    IOCTL_REQUEST(VHOST_VDPA_SET_STATUS);
+    IOCTL_REQUEST(VHOST_SCSI_SET_EVENTS_MISSED);
+    IOCTL_REQUEST(VFIO_IOMMU_DIRTY_PAGES);
+    IOCTL_REQUEST(BTRFS_IOC_SCRUB_PROGRESS);
+    IOCTL_REQUEST(PPPIOCGMRU);
+    IOCTL_REQUEST(BTRFS_IOC_DEV_REPLACE);
+    IOCTL_REQUEST(PPPIOCGFLAGS);
+    IOCTL_REQUEST(NILFS_IOCTL_SET_SUINFO);
+    IOCTL_REQUEST(FW_CDEV_IOC_GET_CYCLE_TIMER2);
+    IOCTL_REQUEST(ATM_DELLECSADDR);
+    IOCTL_REQUEST(FW_CDEV_IOC_GET_SPEED);
+    IOCTL_REQUEST(PPPIOCGIDLE32);
+    IOCTL_REQUEST(VFIO_DEVICE_RESET);
+    IOCTL_REQUEST(GPIO_GET_LINEINFO_UNWATCH_IOCTL);
+    IOCTL_REQUEST(WDIOC_GETSTATUS);
+    IOCTL_REQUEST(BTRFS_IOC_SET_FEATURES);
+    IOCTL_REQUEST(IOCTL_MEI_CONNECT_CLIENT);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_AEWB_CFG);
+    IOCTL_REQUEST(PCITEST_READ);
+    IOCTL_REQUEST(VFIO_GROUP_GET_STATUS);
+    IOCTL_REQUEST(MATROXFB_GET_ALL_OUTPUTS);
+    IOCTL_REQUEST(USBDEVFS_CLEAR_HALT);
+    IOCTL_REQUEST(VIDIOC_DECODER_CMD);
+    IOCTL_REQUEST(VIDIOC_G_AUDIO);
+    IOCTL_REQUEST(CCISS_RESCANDISK);
+    IOCTL_REQUEST(RIO_DISABLE_PORTWRITE_RANGE);
+    IOCTL_REQUEST(IOC_OPAL_SECURE_ERASE_LR);
+    IOCTL_REQUEST(USBDEVFS_REAPURB);
+    IOCTL_REQUEST(DFL_FPGA_CHECK_EXTENSION);
+    IOCTL_REQUEST(AUTOFS_IOC_PROTOVER);
+    IOCTL_REQUEST(FSL_HV_IOCTL_MEMCPY);
+    IOCTL_REQUEST(BTRFS_IOC_GET_FEATURES);
+    IOCTL_REQUEST(PCITEST_MSIX);
+    IOCTL_REQUEST(BTRFS_IOC_DEFRAG_RANGE);
+    IOCTL_REQUEST(UI_BEGIN_FF_ERASE);
+    IOCTL_REQUEST(DM_GET_TARGET_VERSION);
+    IOCTL_REQUEST(PPPIOCGIDLE);
+    IOCTL_REQUEST(NVRAM_SETCKS);
+    IOCTL_REQUEST(WDIOC_GETSUPPORT);
+    IOCTL_REQUEST(GSMIOC_ENABLE_NET);
+    IOCTL_REQUEST(GPIO_GET_CHIPINFO_IOCTL);
+    IOCTL_REQUEST(NE_ADD_VCPU);
+    IOCTL_REQUEST(EVIOCSKEYCODE_V2);
+    IOCTL_REQUEST(PTP_SYS_OFFSET_EXTENDED2);
+    IOCTL_REQUEST(SCIF_FENCE_WAIT);
+    IOCTL_REQUEST(RIO_TRANSFER);
+    IOCTL_REQUEST(FSL_HV_IOCTL_DOORBELL);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_WRITE_LOCAL);
+    IOCTL_REQUEST(I2OEVTREG);
+    IOCTL_REQUEST(I2OPARMGET);
+    IOCTL_REQUEST(EVIOCGID);
+    IOCTL_REQUEST(BTRFS_IOC_QGROUP_CREATE);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_SETPIPEFD);
+    IOCTL_REQUEST(VIDIOC_S_PARM);
+    IOCTL_REQUEST(TUNSETSTEERINGEBPF);
+    IOCTL_REQUEST(ATM_GETNAMES);
+    IOCTL_REQUEST(VIDIOC_QUERYMENU);
+    IOCTL_REQUEST(DFL_FPGA_PORT_DMA_UNMAP);
+    IOCTL_REQUEST(I2OLCTGET);
+    IOCTL_REQUEST(FS_IOC_GET_ENCRYPTION_PWSALT);
+    IOCTL_REQUEST(NS_SETBUFLEV);
+    IOCTL_REQUEST(BLKCLOSEZONE);
+    IOCTL_REQUEST(SONET_GETFRSENSE);
+    IOCTL_REQUEST(UI_SET_EVBIT);
+    IOCTL_REQUEST(DM_LIST_VERSIONS);
+    IOCTL_REQUEST(HIDIOCGSTRING);
+    IOCTL_REQUEST(PPPIOCATTCHAN);
+    IOCTL_REQUEST(VDUSE_DEV_SET_CONFIG);
+    IOCTL_REQUEST(TUNGETFEATURES);
+    IOCTL_REQUEST(VFIO_GROUP_UNSET_CONTAINER);
+    IOCTL_REQUEST(IPMICTL_SET_MY_ADDRESS_CMD);
+    IOCTL_REQUEST(CCISS_REGNEWDISK);
+    IOCTL_REQUEST(VIDIOC_QUERY_DV_TIMINGS);
+    IOCTL_REQUEST(PHN_SETREGS);
+    IOCTL_REQUEST(FAT_IOCTL_GET_ATTRIBUTES);
+#ifdef FSL_MC_SEND_MC_COMMAND
+    IOCTL_REQUEST(FSL_MC_SEND_MC_COMMAND);
+#endif
+    IOCTL_REQUEST(TUNGETIFF);
+    IOCTL_REQUEST(PTP_CLOCK_GETCAPS2);
+#if 0 // needs `struct serial_iso7816`
+    IOCTL_REQUEST(TIOCSISO7816);
+#endif
+    IOCTL_REQUEST(BTRFS_IOC_RESIZE);
+#if 0 // needs `struct usbdevfs_disconnectsignal32`
+    IOCTL_REQUEST(USBDEVFS_DISCSIGNAL32);
+#endif
+    IOCTL_REQUEST(VHOST_SET_VRING_ENDIAN);
+    IOCTL_REQUEST(PPS_KC_BIND);
+    IOCTL_REQUEST(F2FS_IOC_WRITE_CHECKPOINT);
+    IOCTL_REQUEST(UI_SET_FFBIT);
+    IOCTL_REQUEST(IPMICTL_GET_MY_LUN_CMD);
+    IOCTL_REQUEST(CEC_ADAP_G_PHYS_ADDR);
+    IOCTL_REQUEST(CEC_G_MODE);
+    IOCTL_REQUEST(USBDEVFS_RESETEP);
+    IOCTL_REQUEST(MEDIA_REQUEST_IOC_QUEUE);
+    IOCTL_REQUEST(USBDEVFS_ALLOC_STREAMS);
+    IOCTL_REQUEST(MGSL_IOCSXCTRL);
+    IOCTL_REQUEST(MEDIA_IOC_G_TOPOLOGY);
+    IOCTL_REQUEST(PPPIOCUNBRIDGECHAN);
+    IOCTL_REQUEST(F2FS_IOC_COMMIT_ATOMIC_WRITE);
+    IOCTL_REQUEST(ISST_IF_GET_PLATFORM_INFO);
+    IOCTL_REQUEST(SCIF_FENCE_MARK);
+    IOCTL_REQUEST(USBDEVFS_RELEASE_PORT);
+    IOCTL_REQUEST(VFIO_CHECK_EXTENSION);
+    IOCTL_REQUEST(BTRFS_IOC_QGROUP_LIMIT);
+    IOCTL_REQUEST(FAT_IOCTL_GET_VOLUME_ID);
+    IOCTL_REQUEST(UI_SET_PHYS);
+    IOCTL_REQUEST(FDWERRORGET);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_EDID);
+    IOCTL_REQUEST(MGSL_IOCGSTATS);
+    IOCTL_REQUEST(RPROC_SET_SHUTDOWN_ON_RELEASE);
+    IOCTL_REQUEST(SIOCGSTAMP_NEW);
+    IOCTL_REQUEST(RTC_WKALM_RD);
+    IOCTL_REQUEST(PHN_GET_REG);
+    IOCTL_REQUEST(DELL_WMI_SMBIOS_CMD);
+    IOCTL_REQUEST(PHN_NOT_OH);
+    IOCTL_REQUEST(PPGETMODES);
+    IOCTL_REQUEST(CHIOGPARAMS);
+    IOCTL_REQUEST(VFIO_DEVICE_GET_GFX_DMABUF);
+    IOCTL_REQUEST(VHOST_SET_VRING_BUSYLOOP_TIMEOUT);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_G_SELECTION);
+    IOCTL_REQUEST(BTRFS_IOC_RM_DEV_V2);
+    IOCTL_REQUEST(MGSL_IOCWAITGPIO);
+    IOCTL_REQUEST(PMU_IOC_CAN_SLEEP);
+    IOCTL_REQUEST(KCOV_ENABLE);
+    IOCTL_REQUEST(BTRFS_IOC_CLONE);
+    IOCTL_REQUEST(F2FS_IOC_DEFRAGMENT);
+    IOCTL_REQUEST(FW_CDEV_IOC_DEALLOCATE_ISO_RESOURCE);
+    IOCTL_REQUEST(AGPIOC_ALLOCATE);
+    IOCTL_REQUEST(NE_SET_USER_MEMORY_REGION);
+    IOCTL_REQUEST(MGSL_IOCTXABORT);
+    IOCTL_REQUEST(MGSL_IOCSGPIO);
+    IOCTL_REQUEST(LIRC_SET_REC_CARRIER);
+    IOCTL_REQUEST(F2FS_IOC_FLUSH_DEVICE);
+    IOCTL_REQUEST(SNAPSHOT_ATOMIC_RESTORE);
+    IOCTL_REQUEST(RTC_UIE_OFF);
+    IOCTL_REQUEST(BT_BMC_IOCTL_SMS_ATN);
+    IOCTL_REQUEST(NVME_IOCTL_ID);
+    IOCTL_REQUEST(NE_START_ENCLAVE);
+    IOCTL_REQUEST(VIDIOC_STREAMON);
+    IOCTL_REQUEST(FDPOLLDRVSTAT);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_READY);
+    IOCTL_REQUEST(VIDIOC_ENUMAUDOUT);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_S_STD);
+    IOCTL_REQUEST(WDIOC_GETTIMELEFT);
+    IOCTL_REQUEST(ATM_GETLINKRATE);
+    IOCTL_REQUEST(RTC_WKALM_SET);
+    IOCTL_REQUEST(VHOST_GET_BACKEND_FEATURES);
+    IOCTL_REQUEST(ATMARP_ENCAP);
+    IOCTL_REQUEST(CAPI_GET_FLAGS);
+    IOCTL_REQUEST(IPMICTL_SET_MY_CHANNEL_ADDRESS_CMD);
+    IOCTL_REQUEST(DFL_FPGA_FME_PORT_ASSIGN);
+    IOCTL_REQUEST(NS_GET_OWNER_UID);
+    IOCTL_REQUEST(VIDIOC_OVERLAY);
+    IOCTL_REQUEST(BTRFS_IOC_WAIT_SYNC);
+    IOCTL_REQUEST(GPIOHANDLE_SET_CONFIG_IOCTL);
+    IOCTL_REQUEST(VHOST_GET_VRING_ENDIAN);
+    IOCTL_REQUEST(ATM_GETADDR);
+    IOCTL_REQUEST(PHN_GET_REGS);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_REQUESTER);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_EXPIRE);
+    IOCTL_REQUEST(SNAPSHOT_S2RAM);
+    IOCTL_REQUEST(JSIOCSAXMAP);
+    IOCTL_REQUEST(F2FS_IOC_SET_COMPRESS_OPTION);
+    IOCTL_REQUEST(VBG_IOCTL_HGCM_DISCONNECT);
+    IOCTL_REQUEST(SCIF_FENCE_SIGNAL);
+    IOCTL_REQUEST(VFIO_DEVICE_GET_PCI_HOT_RESET_INFO);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_ENUM_MBUS_CODE);
+    IOCTL_REQUEST(MMTIMER_GETOFFSET);
+    IOCTL_REQUEST(RIO_CM_CHAN_LISTEN);
+    IOCTL_REQUEST(ATM_SETSC);
+    IOCTL_REQUEST(F2FS_IOC_SHUTDOWN);
+    IOCTL_REQUEST(NVME_IOCTL_RESCAN);
+    IOCTL_REQUEST(BLKOPENZONE);
+    IOCTL_REQUEST(DM_VERSION);
+    IOCTL_REQUEST(CEC_TRANSMIT);
+    IOCTL_REQUEST(FS_IOC_GET_ENCRYPTION_POLICY_EX);
+    IOCTL_REQUEST(SIOCMKCLIP);
+    IOCTL_REQUEST(IPMI_BMC_IOCTL_CLEAR_SMS_ATN);
+    IOCTL_REQUEST(HIDIOCGVERSION);
+    IOCTL_REQUEST(VIDIOC_S_INPUT);
+    IOCTL_REQUEST(VIDIOC_G_CROP);
+    IOCTL_REQUEST(LIRC_SET_WIDEBAND_RECEIVER);
+    IOCTL_REQUEST(EVIOCGEFFECTS);
+    IOCTL_REQUEST(UVCIOC_CTRL_QUERY);
+    IOCTL_REQUEST(IOC_OPAL_GENERIC_TABLE_RW);
+    IOCTL_REQUEST(FS_IOC_READ_VERITY_METADATA);
+    IOCTL_REQUEST(ND_IOCTL_SET_CONFIG_DATA);
+    IOCTL_REQUEST(USBDEVFS_GETDRIVER);
+    IOCTL_REQUEST(IDT77105_GETSTAT);
+    IOCTL_REQUEST(HIDIOCINITREPORT);
+    IOCTL_REQUEST(VFIO_DEVICE_GET_INFO);
+    IOCTL_REQUEST(RIO_CM_CHAN_RECEIVE);
+    IOCTL_REQUEST(RNDGETENTCNT);
+    IOCTL_REQUEST(PPPIOCNEWUNIT);
+    IOCTL_REQUEST(BTRFS_IOC_INO_LOOKUP);
+    IOCTL_REQUEST(FDRESET);
+    IOCTL_REQUEST(IOC_PR_REGISTER);
+    IOCTL_REQUEST(HIDIOCSREPORT);
+    IOCTL_REQUEST(TEE_IOC_OPEN_SESSION);
+    IOCTL_REQUEST(TEE_IOC_SUPPL_RECV);
+    IOCTL_REQUEST(BTRFS_IOC_BALANCE_CTL);
+    IOCTL_REQUEST(GPIO_GET_LINEINFO_WATCH_IOCTL);
+    IOCTL_REQUEST(HIDIOCGRAWINFO);
+    IOCTL_REQUEST(PPPIOCSCOMPRESS);
+    IOCTL_REQUEST(USBDEVFS_CONNECTINFO);
+    IOCTL_REQUEST(BLKRESETZONE);
+    IOCTL_REQUEST(CHIOINITELEM);
+    IOCTL_REQUEST(NILFS_IOCTL_SET_ALLOC_RANGE);
+    IOCTL_REQUEST(AUTOFS_DEV_IOCTL_CATATONIC);
+    IOCTL_REQUEST(RIO_MPORT_MAINT_HDID_SET);
+    IOCTL_REQUEST(PPGETPHASE);
+    IOCTL_REQUEST(USBDEVFS_DISCONNECT_CLAIM);
+    IOCTL_REQUEST(FDMSGON);
+    IOCTL_REQUEST(VIDIOC_G_SLICED_VBI_CAP);
+    IOCTL_REQUEST(BTRFS_IOC_BALANCE_V2);
+    IOCTL_REQUEST(MEDIA_REQUEST_IOC_REINIT);
+    IOCTL_REQUEST(IOC_OPAL_ERASE_LR);
+    IOCTL_REQUEST(FDFMTBEG);
+    IOCTL_REQUEST(RNDRESEEDCRNG);
+    IOCTL_REQUEST(ISST_IF_GET_PHY_ID);
+    IOCTL_REQUEST(TUNSETNOCSUM);
+    IOCTL_REQUEST(SONET_GETSTAT);
+    IOCTL_REQUEST(TFD_IOC_SET_TICKS);
+    IOCTL_REQUEST(PPDATADIR);
+    IOCTL_REQUEST(IOC_OPAL_ENABLE_DISABLE_MBR);
+    IOCTL_REQUEST(GPIO_V2_GET_LINE_IOCTL);
+    IOCTL_REQUEST(RIO_CM_CHAN_SEND);
+    IOCTL_REQUEST(PPWCTLONIRQ);
+    IOCTL_REQUEST(SONYPI_IOCGBRT);
+    IOCTL_REQUEST(IOC_PR_RELEASE);
+    IOCTL_REQUEST(PPCLRIRQ);
+    IOCTL_REQUEST(IPMICTL_SET_MY_CHANNEL_LUN_CMD);
+    IOCTL_REQUEST(MGSL_IOCSXSYNC);
+    IOCTL_REQUEST(HPET_IE_OFF);
+    IOCTL_REQUEST(IOC_OPAL_ACTIVATE_USR);
+    IOCTL_REQUEST(SONET_SETFRAMING);
+    IOCTL_REQUEST(PERF_EVENT_IOC_PAUSE_OUTPUT);
+    IOCTL_REQUEST(BTRFS_IOC_LOGICAL_INO_V2);
+    IOCTL_REQUEST(VBG_IOCTL_HGCM_CONNECT);
+    IOCTL_REQUEST(BLKFINISHZONE);
+    IOCTL_REQUEST(EVIOCREVOKE);
+    IOCTL_REQUEST(VFIO_DEVICE_FEATURE);
+    IOCTL_REQUEST(CCISS_GETPCIINFO);
+    IOCTL_REQUEST(ISST_IF_MBOX_COMMAND);
+    IOCTL_REQUEST(SCIF_ACCEPTREQ);
+    IOCTL_REQUEST(PERF_EVENT_IOC_QUERY_BPF);
+    IOCTL_REQUEST(VIDIOC_STREAMOFF);
+    IOCTL_REQUEST(VDUSE_DESTROY_DEV);
+    IOCTL_REQUEST(FDGETFDCSTAT);
+    IOCTL_REQUEST(CM_IOCGATR);
+    IOCTL_REQUEST(VIDIOC_S_PRIORITY);
+    IOCTL_REQUEST(SNAPSHOT_FREEZE);
+    IOCTL_REQUEST(VIDIOC_ENUMINPUT);
+    IOCTL_REQUEST(ZATM_GETPOOLZ);
+    IOCTL_REQUEST(RIO_DISABLE_DOORBELL_RANGE);
+    IOCTL_REQUEST(GPIO_V2_GET_LINEINFO_WATCH_IOCTL);
+    IOCTL_REQUEST(VIDIOC_G_STD);
+    IOCTL_REQUEST(USBDEVFS_ALLOW_SUSPEND);
+    IOCTL_REQUEST(SONET_GETSTATZ);
+    IOCTL_REQUEST(SCIF_ACCEPTREG);
+    IOCTL_REQUEST(VIDIOC_ENCODER_CMD);
+    IOCTL_REQUEST(PPPIOCSRASYNCMAP);
+    IOCTL_REQUEST(IOCTL_MEI_NOTIFY_SET);
+    IOCTL_REQUEST(BTRFS_IOC_QUOTA_RESCAN_STATUS);
+    IOCTL_REQUEST(F2FS_IOC_GARBAGE_COLLECT);
+    IOCTL_REQUEST(ATMLEC_CTRL);
+    IOCTL_REQUEST(MATROXFB_GET_AVAILABLE_OUTPUTS);
+    IOCTL_REQUEST(DM_DEV_CREATE);
+    IOCTL_REQUEST(VHOST_VDPA_GET_VRING_NUM);
+    IOCTL_REQUEST(VIDIOC_G_CTRL);
+    IOCTL_REQUEST(NBD_CLEAR_SOCK);
+    IOCTL_REQUEST(VFIO_DEVICE_QUERY_GFX_PLANE);
+    IOCTL_REQUEST(WDIOC_KEEPALIVE);
+    IOCTL_REQUEST(NVME_IOCTL_SUBSYS_RESET);
+    IOCTL_REQUEST(PTP_EXTTS_REQUEST2);
+    IOCTL_REQUEST(PCITEST_BAR);
+    IOCTL_REQUEST(MGSL_IOCGGPIO);
+    IOCTL_REQUEST(EVIOCSREP);
+    IOCTL_REQUEST(VFIO_DEVICE_GET_IRQ_INFO);
+    IOCTL_REQUEST(HPET_DPI);
+    IOCTL_REQUEST(VDUSE_VQ_SETUP_KICKFD);
+    IOCTL_REQUEST(ND_IOCTL_CALL);
+    IOCTL_REQUEST(HIDIOCGDEVINFO);
+    IOCTL_REQUEST(DM_TABLE_DEPS);
+    IOCTL_REQUEST(BTRFS_IOC_DEV_INFO);
+    IOCTL_REQUEST(VDUSE_IOTLB_GET_FD);
+    IOCTL_REQUEST(FW_CDEV_IOC_GET_INFO);
+    IOCTL_REQUEST(VIDIOC_G_PRIORITY);
+    IOCTL_REQUEST(ATM_NEWBACKENDIF);
+    IOCTL_REQUEST(VIDIOC_S_EXT_CTRLS);
+    IOCTL_REQUEST(VIDIOC_SUBDEV_ENUM_DV_TIMINGS);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_CCDC_CFG);
+    IOCTL_REQUEST(VIDIOC_S_HW_FREQ_SEEK);
+    IOCTL_REQUEST(DM_TABLE_LOAD);
+    IOCTL_REQUEST(F2FS_IOC_START_ATOMIC_WRITE);
+    IOCTL_REQUEST(VIDIOC_G_OUTPUT);
+    IOCTL_REQUEST(ATM_DROPPARTY);
+    IOCTL_REQUEST(CHIOGELEM);
+    IOCTL_REQUEST(BTRFS_IOC_GET_SUPPORTED_FEATURES);
+    IOCTL_REQUEST(EVIOCSKEYCODE);
+    IOCTL_REQUEST(NE_GET_IMAGE_LOAD_INFO);
+    IOCTL_REQUEST(TUNSETLINK);
+    IOCTL_REQUEST(FW_CDEV_IOC_ADD_DESCRIPTOR);
+    IOCTL_REQUEST(BTRFS_IOC_SCRUB_CANCEL);
+    IOCTL_REQUEST(PPS_SETPARAMS);
+    IOCTL_REQUEST(IOC_OPAL_LR_SETUP);
+    IOCTL_REQUEST(FW_CDEV_IOC_DEALLOCATE);
+    IOCTL_REQUEST(WDIOC_SETTIMEOUT);
+    IOCTL_REQUEST(IOC_WATCH_QUEUE_SET_FILTER);
+    IOCTL_REQUEST(CAPI_GET_MANUFACTURER);
+    IOCTL_REQUEST(VFIO_IOMMU_SPAPR_UNREGISTER_MEMORY);
+    IOCTL_REQUEST(ASPEED_P2A_CTRL_IOCTL_SET_WINDOW);
+    IOCTL_REQUEST(VIDIOC_G_EDID);
+    IOCTL_REQUEST(F2FS_IOC_GARBAGE_COLLECT_RANGE);
+    IOCTL_REQUEST(RIO_MAP_INBOUND);
+    IOCTL_REQUEST(IOC_OPAL_TAKE_OWNERSHIP);
+    IOCTL_REQUEST(USBDEVFS_CLAIM_PORT);
+    IOCTL_REQUEST(VIDIOC_S_AUDIO);
+    IOCTL_REQUEST(FS_IOC_GET_ENCRYPTION_NONCE);
+    IOCTL_REQUEST(FW_CDEV_IOC_SEND_STREAM_PACKET);
+    IOCTL_REQUEST(BTRFS_IOC_SNAP_DESTROY);
+    IOCTL_REQUEST(SNAPSHOT_FREE);
+    IOCTL_REQUEST(I8K_GET_SPEED);
+    IOCTL_REQUEST(HIDIOCGREPORT);
+    IOCTL_REQUEST(HPET_EPI);
+    IOCTL_REQUEST(JSIOCSCORR);
+    IOCTL_REQUEST(IOC_PR_PREEMPT_ABORT);
+    IOCTL_REQUEST(RIO_MAP_OUTBOUND);
+    IOCTL_REQUEST(ATM_SETESI);
+    IOCTL_REQUEST(FW_CDEV_IOC_START_ISO);
+    IOCTL_REQUEST(ATM_DELADDR);
+    IOCTL_REQUEST(PPFCONTROL);
+    IOCTL_REQUEST(SONYPI_IOCGFAN);
+    IOCTL_REQUEST(RTC_IRQP_SET);
+    IOCTL_REQUEST(PCITEST_WRITE);
+    IOCTL_REQUEST(PPCLAIM);
+    IOCTL_REQUEST(VIDIOC_S_JPEGCOMP);
+    IOCTL_REQUEST(IPMICTL_UNREGISTER_FOR_CMD);
+    IOCTL_REQUEST(VHOST_SET_FEATURES);
+    IOCTL_REQUEST(TOSHIBA_ACPI_SCI);
+    IOCTL_REQUEST(VIDIOC_DQBUF);
+    IOCTL_REQUEST(BTRFS_IOC_BALANCE_PROGRESS);
+    IOCTL_REQUEST(BTRFS_IOC_SUBVOL_SETFLAGS);
+    IOCTL_REQUEST(ATMLEC_MCAST);
+    IOCTL_REQUEST(MMTIMER_GETFREQ);
+    IOCTL_REQUEST(VIDIOC_G_SELECTION);
+    IOCTL_REQUEST(RTC_ALM_SET);
+    IOCTL_REQUEST(PPPOEIOCSFWD);
+    IOCTL_REQUEST(IPMICTL_GET_MAINTENANCE_MODE_CMD);
+    IOCTL_REQUEST(FS_IOC_ENABLE_VERITY);
+    IOCTL_REQUEST(NILFS_IOCTL_GET_BDESCS);
+    IOCTL_REQUEST(FDFMTEND);
+    IOCTL_REQUEST(DMA_BUF_SET_NAME);
+    IOCTL_REQUEST(UI_BEGIN_FF_UPLOAD);
+    IOCTL_REQUEST(RTC_UIE_ON);
+    IOCTL_REQUEST(PPRELEASE);
+    IOCTL_REQUEST(VFIO_IOMMU_UNMAP_DMA);
+    IOCTL_REQUEST(VIDIOC_OMAP3ISP_PRV_CFG);
+    IOCTL_REQUEST(GPIO_GET_LINEHANDLE_IOCTL);
+    IOCTL_REQUEST(VFAT_IOCTL_READDIR_BOTH);
+    IOCTL_REQUEST(NVME_IOCTL_ADMIN_CMD);
+    IOCTL_REQUEST(VHOST_SET_VRING_KICK);
+    IOCTL_REQUEST(BTRFS_IOC_SUBVOL_CREATE_V2);
+    IOCTL_REQUEST(BTRFS_IOC_SNAP_CREATE);
+    IOCTL_REQUEST(SONYPI_IOCGBAT2CAP);
+    IOCTL_REQUEST(PPNEGOT);
+    IOCTL_REQUEST(NBD_PRINT_DEBUG);
+    IOCTL_REQUEST(BTRFS_IOC_INO_LOOKUP_USER);
+    IOCTL_REQUEST(BTRFS_IOC_GET_SUBVOL_ROOTREF);
+    IOCTL_REQUEST(FS_IOC_REMOVE_ENCRYPTION_KEY_ALL_USERS);
+    IOCTL_REQUEST(BTRFS_IOC_FS_INFO);
+    IOCTL_REQUEST(VIDIOC_ENUM_FMT);
+    IOCTL_REQUEST(VIDIOC_G_INPUT);
+    IOCTL_REQUEST(VTPM_PROXY_IOC_NEW_DEV);
+    IOCTL_REQUEST(DFL_FPGA_FME_ERR_GET_IRQ_NUM);
+    IOCTL_REQUEST(ND_IOCTL_DIMM_FLAGS);
+    IOCTL_REQUEST(BTRFS_IOC_QUOTA_RESCAN);
+    IOCTL_REQUEST(MMTIMER_GETCOUNTER);
+    IOCTL_REQUEST(MATROXFB_GET_OUTPUT_MODE);
+    IOCTL_REQUEST(BTRFS_IOC_QUOTA_RESCAN_WAIT);
+    IOCTL_REQUEST(RIO_CM_CHAN_BIND);
+    IOCTL_REQUEST(HIDIOCGRDESC);
+    IOCTL_REQUEST(MGSL_IOCGIF);
+    IOCTL_REQUEST(VIDIOC_S_OUTPUT);
+    IOCTL_REQUEST(HIDIOCGREPORTINFO);
+    IOCTL_REQUEST(WDIOC_GETBOOTSTATUS);
+    IOCTL_REQUEST(VDUSE_VQ_GET_INFO);
+    IOCTL_REQUEST(ACRN_IOCTL_ASSIGN_PCIDEV);
+#ifdef BLKGETDISKSEQ
+    IOCTL_REQUEST(BLKGETDISKSEQ);
+#endif
+    IOCTL_REQUEST(ACRN_IOCTL_PM_GET_CPU_STATE);
+    IOCTL_REQUEST(ACRN_IOCTL_DESTROY_VM);
+    IOCTL_REQUEST(ACRN_IOCTL_SET_PTDEV_INTR);
+    IOCTL_REQUEST(ACRN_IOCTL_CREATE_IOREQ_CLIENT);
+    IOCTL_REQUEST(ACRN_IOCTL_IRQFD);
+    IOCTL_REQUEST(ACRN_IOCTL_CREATE_VM);
+    IOCTL_REQUEST(ACRN_IOCTL_INJECT_MSI);
+    IOCTL_REQUEST(ACRN_IOCTL_ATTACH_IOREQ_CLIENT);
+    IOCTL_REQUEST(ACRN_IOCTL_RESET_PTDEV_INTR);
+    IOCTL_REQUEST(ACRN_IOCTL_NOTIFY_REQUEST_FINISH);
+    IOCTL_REQUEST(ACRN_IOCTL_SET_IRQLINE);
+    IOCTL_REQUEST(ACRN_IOCTL_START_VM);
+    IOCTL_REQUEST(ACRN_IOCTL_SET_VCPU_REGS);
+    IOCTL_REQUEST(ACRN_IOCTL_SET_MEMSEG);
+    IOCTL_REQUEST(ACRN_IOCTL_PAUSE_VM);
+    IOCTL_REQUEST(ACRN_IOCTL_CLEAR_VM_IOREQ);
+    IOCTL_REQUEST(ACRN_IOCTL_UNSET_MEMSEG);
+    IOCTL_REQUEST(ACRN_IOCTL_IOEVENTFD);
+    IOCTL_REQUEST(ACRN_IOCTL_DEASSIGN_PCIDEV);
+    IOCTL_REQUEST(ACRN_IOCTL_RESET_VM);
+    IOCTL_REQUEST(ACRN_IOCTL_DESTROY_IOREQ_CLIENT);
+    IOCTL_REQUEST(ACRN_IOCTL_VM_INTR_MONITOR);
+}
