+++ src/mips/general.rs
@@ -1826,26 +1826,6 @@ pub const TCION: u32 = 3;
 pub const TCIFLUSH: u32 = 0;
 pub const TCOFLUSH: u32 = 1;
 pub const TCIOFLUSH: u32 = 2;
-pub const TCGETA: u32 = 21505;
-pub const TCSETA: u32 = 21506;
-pub const TCSETAW: u32 = 21507;
-pub const TCSETAF: u32 = 21508;
-pub const TCSBRK: u32 = 21509;
-pub const TCXONC: u32 = 21510;
-pub const TCFLSH: u32 = 21511;
-pub const TCGETS: u32 = 21517;
-pub const TCSETS: u32 = 21518;
-pub const TCSETSW: u32 = 21519;
-pub const TCSETSF: u32 = 21520;
-pub const TIOCEXCL: u32 = 29709;
-pub const TIOCNXCL: u32 = 29710;
-pub const TIOCOUTQ: u32 = 29810;
-pub const TIOCSTI: u32 = 21618;
-pub const TIOCMGET: u32 = 29725;
-pub const TIOCMBIS: u32 = 29723;
-pub const TIOCMBIC: u32 = 29724;
-pub const TIOCMSET: u32 = 29722;
-pub const TIOCPKT: u32 = 21616;
 pub const TIOCPKT_DATA: u32 = 0;
 pub const TIOCPKT_FLUSHREAD: u32 = 1;
 pub const TIOCPKT_FLUSHWRITE: u32 = 2;
@@ -1854,43 +1834,11 @@ pub const TIOCPKT_START: u32 = 8;
 pub const TIOCPKT_NOSTOP: u32 = 16;
 pub const TIOCPKT_DOSTOP: u32 = 32;
 pub const TIOCPKT_IOCTL: u32 = 64;
-pub const TIOCNOTTY: u32 = 21617;
-pub const TIOCSETD: u32 = 29697;
-pub const TIOCGETD: u32 = 29696;
-pub const FIOCLEX: u32 = 26113;
-pub const FIONCLEX: u32 = 26114;
-pub const FIOASYNC: u32 = 26237;
-pub const FIONBIO: u32 = 26238;
-pub const FIOQSIZE: u32 = 26239;
 pub const TIOCGLTC: u32 = 29812;
 pub const TIOCSLTC: u32 = 29813;
-pub const FIONREAD: u32 = 18047;
-pub const TIOCINQ: u32 = 18047;
 pub const TIOCGETP: u32 = 29704;
 pub const TIOCSETP: u32 = 29705;
 pub const TIOCSETN: u32 = 29706;
-pub const TIOCSBRK: u32 = 21543;
-pub const TIOCCBRK: u32 = 21544;
-pub const TIOCGSID: u32 = 29718;
-pub const TIOCVHANGUP: u32 = 21559;
-pub const TIOCSCTTY: u32 = 21632;
-pub const TIOCGSOFTCAR: u32 = 21633;
-pub const TIOCSSOFTCAR: u32 = 21634;
-pub const TIOCLINUX: u32 = 21635;
-pub const TIOCGSERIAL: u32 = 21636;
-pub const TIOCSSERIAL: u32 = 21637;
-pub const TCSBRKP: u32 = 21638;
-pub const TIOCSERCONFIG: u32 = 21640;
-pub const TIOCSERGWILD: u32 = 21641;
-pub const TIOCSERSWILD: u32 = 21642;
-pub const TIOCGLCKTRMIOS: u32 = 21643;
-pub const TIOCSLCKTRMIOS: u32 = 21644;
-pub const TIOCSERGSTRUCT: u32 = 21645;
-pub const TIOCSERGETLSR: u32 = 21646;
-pub const TIOCSERGETMULTI: u32 = 21647;
-pub const TIOCSERSETMULTI: u32 = 21648;
-pub const TIOCMIWAIT: u32 = 21649;
-pub const TIOCGICOUNT: u32 = 21650;
 pub const NCC: u32 = 8;
 pub const TIOCM_LE: u32 = 1;
 pub const TIOCM_DTR: u32 = 2;
