+++ src/powerpc64/general.rs
@@ -978,11 +978,6 @@ pub const SO_RCVTIMEO_OLD: u32 = 18;
 pub const SO_SNDTIMEO_OLD: u32 = 19;
 pub const SO_PASSCRED: u32 = 20;
 pub const SO_PEERCRED: u32 = 21;
-pub const FIOSETOWN: u32 = 35073;
-pub const SIOCSPGRP: u32 = 35074;
-pub const FIOGETOWN: u32 = 35075;
-pub const SIOCGPGRP: u32 = 35076;
-pub const SIOCATMARK: u32 = 35077;
 pub const SIOCGSTAMP_OLD: u32 = 35078;
 pub const SIOCGSTAMPNS_OLD: u32 = 35079;
 pub const SOL_SOCKET: u32 = 1;
@@ -1571,14 +1566,6 @@ pub const TCP_MD5SIG_MAXKEYLEN: u32 = 80;
 pub const TCP_MD5SIG_FLAG_PREFIX: u32 = 1;
 pub const TCP_MD5SIG_FLAG_IFINDEX: u32 = 2;
 pub const TCP_RECEIVE_ZEROCOPY_FLAG_TLB_CLEAN_HINT: u32 = 1;
-pub const TIOCEXCL: u32 = 21516;
-pub const TIOCNXCL: u32 = 21517;
-pub const TIOCSCTTY: u32 = 21518;
-pub const TIOCSTI: u32 = 21522;
-pub const TIOCMGET: u32 = 21525;
-pub const TIOCMBIS: u32 = 21526;
-pub const TIOCMBIC: u32 = 21527;
-pub const TIOCMSET: u32 = 21528;
 pub const TIOCM_LE: u32 = 1;
 pub const TIOCM_DTR: u32 = 2;
 pub const TIOCM_RTS: u32 = 4;
@@ -1593,13 +1580,6 @@ pub const TIOCM_RI: u32 = 128;
 pub const TIOCM_OUT1: u32 = 8192;
 pub const TIOCM_OUT2: u32 = 16384;
 pub const TIOCM_LOOP: u32 = 32768;
-pub const TIOCGSOFTCAR: u32 = 21529;
-pub const TIOCSSOFTCAR: u32 = 21530;
-pub const TIOCLINUX: u32 = 21532;
-pub const TIOCCONS: u32 = 21533;
-pub const TIOCGSERIAL: u32 = 21534;
-pub const TIOCSSERIAL: u32 = 21535;
-pub const TIOCPKT: u32 = 21536;
 pub const TIOCPKT_DATA: u32 = 0;
 pub const TIOCPKT_FLUSHREAD: u32 = 1;
 pub const TIOCPKT_FLUSHWRITE: u32 = 2;
@@ -1608,28 +1588,7 @@ pub const TIOCPKT_START: u32 = 8;
 pub const TIOCPKT_NOSTOP: u32 = 16;
 pub const TIOCPKT_DOSTOP: u32 = 32;
 pub const TIOCPKT_IOCTL: u32 = 64;
-pub const TIOCNOTTY: u32 = 21538;
-pub const TIOCSETD: u32 = 21539;
-pub const TIOCGETD: u32 = 21540;
-pub const TCSBRKP: u32 = 21541;
-pub const TIOCSBRK: u32 = 21543;
-pub const TIOCCBRK: u32 = 21544;
-pub const TIOCGSID: u32 = 21545;
-pub const TIOCGRS485: u32 = 21550;
-pub const TIOCSRS485: u32 = 21551;
-pub const TIOCVHANGUP: u32 = 21559;
-pub const TIOCSERCONFIG: u32 = 21587;
-pub const TIOCSERGWILD: u32 = 21588;
-pub const TIOCSERSWILD: u32 = 21589;
-pub const TIOCGLCKTRMIOS: u32 = 21590;
-pub const TIOCSLCKTRMIOS: u32 = 21591;
-pub const TIOCSERGSTRUCT: u32 = 21592;
-pub const TIOCSERGETLSR: u32 = 21593;
 pub const TIOCSER_TEMT: u32 = 1;
-pub const TIOCSERGETMULTI: u32 = 21594;
-pub const TIOCSERSETMULTI: u32 = 21595;
-pub const TIOCMIWAIT: u32 = 21596;
-pub const TIOCGICOUNT: u32 = 21597;
 pub const NCCS: u32 = 19;
 pub const VINTR: u32 = 0;
 pub const VQUIT: u32 = 1;
