+++ src/riscv32/general.rs
@@ -962,11 +962,6 @@ pub const MAP_HUGE_512MB: u32 = 1946157056;
 pub const MAP_HUGE_1GB: u32 = 2013265920;
 pub const MAP_HUGE_2GB: u32 = 2080374784;
 pub const MAP_HUGE_16GB: u32 = 2281701376;
-pub const FIOSETOWN: u32 = 35073;
-pub const SIOCSPGRP: u32 = 35074;
-pub const FIOGETOWN: u32 = 35075;
-pub const SIOCGPGRP: u32 = 35076;
-pub const SIOCATMARK: u32 = 35077;
 pub const SIOCGSTAMP_OLD: u32 = 35078;
 pub const SIOCGSTAMPNS_OLD: u32 = 35079;
 pub const SOL_SOCKET: u32 = 1;
@@ -1691,69 +1686,6 @@ pub const TCIOFLUSH: u32 = 2;
 pub const TCSANOW: u32 = 0;
 pub const TCSADRAIN: u32 = 1;
 pub const TCSAFLUSH: u32 = 2;
-pub const TCGETS: u32 = 21505;
-pub const TCSETS: u32 = 21506;
-pub const TCSETSW: u32 = 21507;
-pub const TCSETSF: u32 = 21508;
-pub const TCGETA: u32 = 21509;
-pub const TCSETA: u32 = 21510;
-pub const TCSETAW: u32 = 21511;
-pub const TCSETAF: u32 = 21512;
-pub const TCSBRK: u32 = 21513;
-pub const TCXONC: u32 = 21514;
-pub const TCFLSH: u32 = 21515;
-pub const TIOCEXCL: u32 = 21516;
-pub const TIOCNXCL: u32 = 21517;
-pub const TIOCSCTTY: u32 = 21518;
-pub const TIOCGPGRP: u32 = 21519;
-pub const TIOCSPGRP: u32 = 21520;
-pub const TIOCOUTQ: u32 = 21521;
-pub const TIOCSTI: u32 = 21522;
-pub const TIOCGWINSZ: u32 = 21523;
-pub const TIOCSWINSZ: u32 = 21524;
-pub const TIOCMGET: u32 = 21525;
-pub const TIOCMBIS: u32 = 21526;
-pub const TIOCMBIC: u32 = 21527;
-pub const TIOCMSET: u32 = 21528;
-pub const TIOCGSOFTCAR: u32 = 21529;
-pub const TIOCSSOFTCAR: u32 = 21530;
-pub const FIONREAD: u32 = 21531;
-pub const TIOCINQ: u32 = 21531;
-pub const TIOCLINUX: u32 = 21532;
-pub const TIOCCONS: u32 = 21533;
-pub const TIOCGSERIAL: u32 = 21534;
-pub const TIOCSSERIAL: u32 = 21535;
-pub const TIOCPKT: u32 = 21536;
-pub const FIONBIO: u32 = 21537;
-pub const TIOCNOTTY: u32 = 21538;
-pub const TIOCSETD: u32 = 21539;
-pub const TIOCGETD: u32 = 21540;
-pub const TCSBRKP: u32 = 21541;
-pub const TIOCSBRK: u32 = 21543;
-pub const TIOCCBRK: u32 = 21544;
-pub const TIOCGSID: u32 = 21545;
-pub const TIOCGRS485: u32 = 21550;
-pub const TIOCSRS485: u32 = 21551;
-pub const TCGETX: u32 = 21554;
-pub const TCSETX: u32 = 21555;
-pub const TCSETXF: u32 = 21556;
-pub const TCSETXW: u32 = 21557;
-pub const TIOCVHANGUP: u32 = 21559;
-pub const FIONCLEX: u32 = 21584;
-pub const FIOCLEX: u32 = 21585;
-pub const FIOASYNC: u32 = 21586;
-pub const TIOCSERCONFIG: u32 = 21587;
-pub const TIOCSERGWILD: u32 = 21588;
-pub const TIOCSERSWILD: u32 = 21589;
-pub const TIOCGLCKTRMIOS: u32 = 21590;
-pub const TIOCSLCKTRMIOS: u32 = 21591;
-pub const TIOCSERGSTRUCT: u32 = 21592;
-pub const TIOCSERGETLSR: u32 = 21593;
-pub const TIOCSERGETMULTI: u32 = 21594;
-pub const TIOCSERSETMULTI: u32 = 21595;
-pub const TIOCMIWAIT: u32 = 21596;
-pub const TIOCGICOUNT: u32 = 21597;
-pub const FIOQSIZE: u32 = 21600;
 pub const TIOCPKT_DATA: u32 = 0;
 pub const TIOCPKT_FLUSHREAD: u32 = 1;
 pub const TIOCPKT_FLUSHWRITE: u32 = 2;
